% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Census_state_population_M3T}
\alias{Census_state_population_M3T}
\title{State population data from the U.S. Census Bureau}
\format{
\code{Census_state_population_M3T} A data frame with 49 rows and 16 columns:
\describe{
\item{Name}{State name}
\item{POPESTIMATE2010}{Estimated population for the year 2010}
\item{POPESTIMATE2011}{Estimated population for the year 2011}
\item{POPESTIMATE2012}{Estimated population for the year 2012}
\item{POPESTIMATE2013}{Estimated population for the year 2013}
\item{POPESTIMATE2014}{Estimated population for the year 2014}
\item{POPESTIMATE2015}{Estimated population for the year 2015}
\item{POPESTIMATE2016}{Estimated population for the year 2016}
\item{POPESTIMATE2017}{Estimated population for the year 2017}
\item{POPESTIMATE2018}{Estimated population for the year 2018}
\item{POPESTIMATE2019}{Estimated population for the year 2019}
\item{POPESTIMATE2020}{Estimated population for the year 2020}
\item{POPESTIMATE2021}{Estimated population for the year 2021}
\item{POPESTIMATE2022}{Estimated population for the year 2022}
\item{POPESTIMATE2023}{Estimated population for the year 2023}
\item{POPESTIMATE2024}{Estimated population for the year 2024}
}
}
\source{
\url{https://www.census.gov/data/tables/time-series/demo/popest/2020s-state-total.html}
}
\usage{
Census_state_population_M3T
}
\description{
A subset of data from the Census Bureau
}
\details{
The estimated state level population published by the U.S. Census Bureau
}
\keyword{datasets}
