% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH4_inventory_build.R
\name{CH4_inventory_build}
\alias{CH4_inventory_build}
\title{Main function to calculate gridded methane emissions.  Runs sector by
sector according to the config file.}
\usage{
CH4_inventory_build(
  run_directory,
  inventory_year,
  domain = "custom",
  domain_res = 0.01,
  domain_crs = "epsg:4326",
  verbose = FALSE,
  EIA_file = "default",
  HIFLD_compressor_file = "default",
  DMR_file = "default",
  CWNS_file = "default",
  PHMSA_file = "default",
  Wetcharts_file = "default"
)
}
\arguments{
\item{run_directory}{Character providing the full filepath to load/save
input and output data.  Subfolders will be created.}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used for datasets that are not provided annually
with a user update.}

\item{domain}{data.frame or character.  If data.frame, provides the corner
coordinates of the desired region to process.  The first column would be x
values, the second column y values, and the first row would be the minima,
the second row the maxima.  These must be provided in the appropriate units
for the domain_crs parameter (e.g., decimal degrees for a lat/long
projection).

If character, it can be based on census Tigerlines.  It can be a state
abbreviation, state name, state FIPS code (as a character), Urban Area name,
or Urban Area Census Code.  For example: "DE", "Delaware", and "10" are all
equivalent and "Long Neck, DE" and "51202" are equivalent. It can also be
"CONUS" for the entire continental United States or "custom" for a domain
created by the user interactively on a US map. Lastly, it can be a filepath
pointing to a polygon file that can be interpreted by the terra package.
Names must match exactly and FIPS or urban area codes must include leading
0's.  Lists of the names and codes for all urban areas for the most recent
census are available at
\url{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural.html}.
Links to previous census urban areas are on the same page.  A list with
state fips codes is available here
\url{https://www.census.gov/library/reference/code-lists/ansi.html}.}

\item{domain_res}{Numeric providing the resolution for the domain.  Can be
length 1 for equal x and y resolution or length 2 (x, y).}

\item{domain_crs}{Character providing the projection of the domain in PROJ
string (\url{https://proj.org/en/stable/operations/projections/index.html})
or EPSG codes (\url{https://epsg.io/}) or WKT
(\url{https://docs.ogc.org/is/12-063r5/12-063r5.html}) formats.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes easier to read csv files of the data that has been gridded for
multiple sectors as well as sector and subsector visuals.}

\item{EIA_file}{Character providing the full filepath to the EIA form 176 data
in xlsx format available at
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}
or "default" to use assets built into the package or it's companion Zenodo.}

\item{HIFLD_compressor_file}{Character providing the full filepath to the
HIFLD natural gas transmission compressor data in csv format or "default" to
use assets built into the package or it's companion Zenodo.  This data has
been deprecated and is no longer available from HIFLD.}

\item{DMR_file}{Character providing the full filepath to the DMR wastewater
treatment plant data in csv format available at
\url{https://echo.epa.gov/trends/loading-tool/water-pollution-search} or
"default" to use assets built into the package or it's companion Zenodo.}

\item{CWNS_file}{Character providing the full filepath to the CWNS wastewater
treatment plant data or "default" to use assets built into the package or
it's companion Zenodo.  For 2012 this should be in xlsx format and for 2022
this should be a folder with multiple csv's. The 2012 CWNS is available at
\url{https://ordspub.epa.gov/ords/cwns2012/f?p=cwns2012:25} and the 2022
CWNS is available at
\url{https://sdwis.epa.gov/ords/sfdw_pub/r/sfdw/cwns_pub/data-download?session=8491118346746}.}

\item{PHMSA_file}{Character providing the full filepath to the PHMSA Gas
Distribution Annual Data xlsx file available at
\url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
or "default" to use assets built into the package or it's companion Zenodo.
Zip files for groups of years are available via links at the bottom of the
page.  The default cannot be used to run at the LDC level.}

\item{Wetcharts_file}{Character providing the full filepath to the Wetcharts
file to be used in .nc format available at
\url{https://doi.org/10.3334/ORNLDAAC/2346} or "default" to use assets built
into the package or it's companion Zenodo.}
}
\value{
Nothing is returned from the function, but there will be frequent
user updates as multiple other functions are run and output files created
for the various sectors.
}
\description{
\code{CH4_inventory_build} runs multiple other functions to calculate
the gridded methane emissions sector-by-sector using information provided in
the config file and inputs.
}
\details{
This function will call all other functions and the config, create
the domain from the input data, and download some files (Census tigerlines,
ghgrp location/name data) that will be needed for multiple sectors.  It will
then use the information in the config file and inputs to run every relevant
sector.  As such, the inputs to this function and the config may require
user editing.

See references \href{https://doi.org/10.1029/2020JD032974}{Vulcan} and
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\examples{
CH4_inventory_build(run_directory="~/../Desktop/methane_inventory_2025_NE_run/",
                    inventory_year=2019,
                    domain=as.data.frame(cbind(c(-76.65,-73.65),
                                               c(38.97,40.97))),
                    domain_res=0.01,
                    domain_crs="epsg:4326",
                    verbose=TRUE)
}
\seealso{
\code{\link[=Set_config]{Set_config()}} Generates the config function with user-editable
settings used throughout processing.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
