% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combiner.R
\name{Combine_inventories}
\alias{Combine_inventories}
\title{Create gridded methane emissions maps that include all sectors}
\usage{
Combine_inventories(
  output_directory,
  separate_thermo,
  plot_directory,
  County_Tigerlines,
  State_CB,
  domain,
  domain_template,
  verbose
)
}
\arguments{
\item{output_directory}{Character providing the full filepath to save
processed data}

\item{separate_thermo}{Logical.  Pulled from config file.  Indicating whether
or not to save partial inventories: one that is only thermogenic sources,
and one that is only non-thermogenic sources.}
}
\value{
Nothing is returned from the function, but the main outputs are many
netcdf files of the methane emissions across all sectors.  Given the large
number of possible files and variations, they are titled
"Combined_inventory_combination_1.nc" increasing numerically.  A csv is also
saved that details what variations were used for each sector for each
inventory.
}
\description{
\code{Combine_inventories} writes a netcdf for every unique
combination of sectoral emissions estimates.  Optionally writes additional
netcdf with the thermogenic and nonthermogenic sectors separately.
}
\details{
This function considers every possible output across all sectors and
filters to keep those that have been saved to the output folder.  From there
it creates a table of every possible unique combination of sectors.  If many
options have been enabled, this can be over 1,000 combinations, so caution
is warranted, though processing is generally fast and file sizes small -
this will depend strongly on the domain size. Additionally, if the config
option is set, three inventories will be created for each unique combination
instead:
\itemize{
\item One that is the total across all sectors
\item One that is the total across only thermogenic sources
\itemize{
\item stationary combustion of wood
\item natural gas distribution
\item natural gas transmission
\item Gridded EPA mobile combustion
\item Gridded EPA abandoned coal
\item Gridded EPA surface coal
\item Gridded EPA underground coal
\item Gridded EPA petroleum systems exploration
\item Gridded EPA petroleum systems production
\item Gridded EPA petroleum systems refining
\item Gridded EPA petroleum systems transport
\item Gridded EPA abandoned oil gas
\item Gridded EPA natural gas exploration
\item Gridded EPA natural gas processing
\item Gridded EPA natural gas production
\item Gridded EPA industry petrochemical
\item Gridded EPA industry ferroalloy
}
\item One that is the total across only non-thermogenic sources
\itemize{
\item landfills
\item stationary combustion of fossil fuels
\item wastewater
\item wetlands
\item Gridded EPA composting
\item Gridded EPA enteric fermentation
\item Gridded EPA manure management
\item Gridded EPA rice cultivation
\item Gridded EPA field burning
}
}
}
\examples{
library(terra)
Combine_inventories <- function(output_directory="~/../Desktop/out/",
                               separate_thermo=T)
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
