% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CWNS_2012}
\alias{CWNS_2012}
\title{2012 Clean Watershed Needs Survey (CWNS) data on waste water treatment plants}
\format{
\subsection{\code{CWNS_2012} A data frame with 13,360 rows and 4 columns:}{

\describe{
\item{LATITUDE}{Facility location coordinates}
\item{LONGITUDE}{Facility location coordinates}
\item{EXIST_MUNICIPAL}{Annual flow of municipal wastewater through a facility in millions of gallons per day}
\item{HORIZONTAL_COORDINATE_DATUM}{Information regarding the mapping used in the coordinates}
}
}
}
\source{
\url{https://www.epa.gov/cwns/clean-watersheds-needs-survey-cwns-2012-report-and-data}
}
\usage{
CWNS_2012
}
\description{
A subset of data from the 2012 CWNS
}
\details{
The 2012 CWNS is available as an access database.  The relevant data was
pulled and the file was converted to a csv file as described in detail in
\code{\link{M3T_config}}.
}
\keyword{datasets}
