% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CWNS_2022}
\alias{CWNS_2022}
\title{2022 Clean Watershed Needs Survey (CWNS) data on waste water treatment plants}
\format{
\subsection{\code{CWNS_2022} A data frame with 13,270 rows and 3 columns:}{

\describe{
\item{LATITUDE}{Facility location coordinates}
\item{LONGITUDE}{Facility location coordinates}
\item{EXIST_MUNICIPAL}{Annual flow of municipal wastewater through a facility in millions of gallons per day}
}
}
}
\source{
\url{https://www.epa.gov/cwns/clean-watersheds-needs-survey-cwns-2022-report-and-data}
}
\usage{
CWNS_2022
}
\description{
A subset of data from the 2022 CWNS
}
\details{
The 2022 CWNS is available as several CSVs by going to the data dashboard and
selecting data download in the top right.  The relevant data was pulled from
the various files, combined, and saved to a single CSV.
}
\keyword{datasets}
