% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_custom_domain.R
\name{define_custom_domain}
\alias{define_custom_domain}
\title{Interactively define the domain for 'CH4_inventory_build'}
\usage{
define_custom_domain(input_directory, State_CB)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{State_CB}{SpatVector.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization.}
}
\value{
SpatVector
}
\description{
This function uses the terra package to interactively build a
domain area using a state map of the continental US.
}
\details{
The polygon is saved for potential reuse and reference.  Be aware
that nonsensical polygons can be built (e.g., those with holes, odd shapes,
etc.) so the user check at the end is quite important to make sure this
polygon is useful.
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.
}
\keyword{internal}
