% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EIA_NG_data}
\alias{EIA_NG_data}
\title{Data from EIA Form 176 - the Annual Report of Natural and supplemental Gas
Supply and Disposition}
\format{
\code{EIA_NG_data} A data frame with 28,553 rows and 10 columns:
\describe{
\item{Year}{Year of data}
\item{State}{State the utility operates in}
\item{Residential_Total_Volume_(Mcf)}{Volume of natural gas delivered to residential customers in thousand cubic feet}
\item{Residential_Total_Customers}{Number of residential customers}
\item{Commercial_Total_Volume_(Mcf)}{Volume of natural gas delivered to commercial customers in thousand cubic feet}
\item{Commercial_Total_Customers}{Number of commercial customers}
\item{Industrial_Total_Volume_(Mcf)}{Volume of natural gas delivered to industrial customers in thousand cubic feet}
\item{Industrial_Total_Customers}{Number of industrial customers}
\item{Electric_Total_Volume_(Mcf)}{Volume of natural gas delivered to electric production customers in thousand cubic feet}
\item{Electric_Total_Customers}{Number of electric production customers}
}
}
\source{
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}
}
\usage{
EIA_NG_data
}
\description{
A subset of data from EIA Form 176
}
\details{
The volume of natural gas delivered and number of customers for natural gas
utilities, separated by customer type.  The data can be downloaded as a csv
}
\keyword{datasets}
