% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EIA_SEDS}
\alias{EIA_SEDS}
\title{Data from the EIA State Energy Data System (SEDS) for all continental US
states and all relevant years available at the time}
\format{
\code{EIA_SEDS} A data frame with 8,918 rows and 5 columns:
\describe{
\item{period}{Year of data}
\item{seriesId}{ID code of the fuel and sector}
\item{seriesDescription}{description of the seriesId}
\item{stateId}{State abbreviation for the data}
\item{value}{Consumed fuel for the corresponding state, year, fuel, and sector in billions of BTU}
}
}
\source{
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}
}
\usage{
EIA_SEDS
}
\description{
A subset of data from the EIA SEDS
}
\details{
The amount of fuel consumed separated by fuel, state, year, and economic
sector
}
\keyword{datasets}
