% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGI_landfill_total}
\alias{GHGI_landfill_total}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI)}
\format{
\code{GHGI_landfill_total} A data frame with 12 rows and 2 columns:
\describe{
\item{Year}{Year of data}
\item{Emissions}{National total emissions in kilotons/year for municipal solid waste as reported in the GHGI}
}
}
\source{
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022}
}
\usage{
GHGI_landfill_total
}
\description{
A subset of data from the GHGI
}
\details{
The national methane emissions estimated for municipal landfills in the GHGI
}
\keyword{datasets}
