% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGI_NG_distribution}
\alias{GHGI_NG_distribution}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI)}
\format{
\code{GHGI_NG_distribution} A list of data frames:
\describe{
\item{GHGI_MnR_EF}{Meter and regulating station emission factors in mol/s with columns representing years and rows representing type}
\item{GHGI_MnR_Activity}{Meter and regulating station counts with columns representing years and rows representing type}
\item{GHGI_services}{Service line emission factors in mol/s with columns representing years and rows representing material}
\item{GHGI_meters}{Meter emission factors in mol/s with columns representing years and rows representing customer type}
\item{GHGI_maintenance}{Maintenance/mishap emission factors in mol/s with columns representing years and rows representing type}
}
}
\source{
\url{https://www.epa.gov/ghgemissions/natural-gas-and-petroleum-systems-ghg-inventory-additional-information-1990-2021-ghg}
}
\usage{
GHGI_NG_distribution
}
\description{
A subset of data from the GHGI
}
\details{
The national methane emission factors and activity data estimated for natural
gas distribution in the GHGI
}
\keyword{datasets}
