% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGI_NG_transmission}
\alias{GHGI_NG_transmission}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI)}
\format{
\code{GHGI_NG_transmission} A list of data frames:
\describe{
\item{GHGI_Pipeline_Activity}{Pipeline equipment count or length of pipeline in meters with columns representing years and rows representing type}
\item{GHGI_Pipeline_Emissions}{Pipeline equipment emissions in mol/s with columns representing years and rows representing type}
\item{GHGI_transmission_compressors_Activity}{Transmission compressor counts with columns representing years and rows representing material}
\item{GHGI_transmission_compressors_Emissions}{Transmission compressor emissions in mol/s with columns representing years and rows representing type}
}
}
\source{
\url{https://www.epa.gov/ghgemissions/natural-gas-and-petroleum-systems-ghg-inventory-additional-information-1990-2021-ghg}
}
\usage{
GHGI_NG_transmission
}
\description{
A subset of data from the GHGI
}
\details{
The national methane emission factors and activity data estimated for natural
gas distribution in the GHGI
}
\keyword{datasets}
