% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGI_stationary_combustion}
\alias{GHGI_stationary_combustion}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI)}
\format{
\code{GHGI_stationary_combustion} A data frame with data year as
rownames, 12 rows, and 15 columns:
\describe{
\item{State}{US_EPA for all entries}
\item{com_coal}{national coal use by the commercial sector in trillions of BTU}
\item{ind_coal}{national coal use by the industrial sector in trillions of BTU}
\item{elec_coal}{national coal use by the electric production sector in trillions of BTU}
\item{res_petr}{national petroleum use by the residential sector in trillions of BTU}
\item{com_petr}{national petroleum use by the commercial sector in trillions of BTU}
\item{ind_petr}{national petroleum use by the industrial sector in trillions of BTU}
\item{elec_petr}{national petroleum use by the electric production sector in trillions of BTU}
\item{com_gas}{national natural gas use by the commercial sector in trillions of BTU}
\item{ind_gas}{national natural gas use by the industrial sector in trillions of BTU}
\item{elec_gas}{national natural gas use by the electric production sector in trillions of BTU}
\item{res_wood}{national wood use by the residential sector in trillions of BTU}
\item{com_wood}{national wood use by the commercial sector in trillions of BTU}
\item{ind_wood}{national wood use by the industrial sector in trillions of BTU}
\item{elec_wood}{national wood use by the electric production sector in trillions of BTU}
}
}
\source{
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022}
}
\usage{
GHGI_stationary_combustion
}
\description{
A subset of data from the GHGI
}
\details{
The national fuel consumption data estimated in the GHGI
}
\keyword{datasets}
