% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGRP_combustion}
\alias{GHGRP_combustion}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Reporting Program (GHGRP)}
\format{
\code{GHGRP_combustion} A data frame with 68,122 rows, and 5 columns:
\describe{
\item{facility_id}{Unique GHGRP ID for each facility}
\item{facility_name}{Name of facility}
\item{ghg_name}{greenhouse gas name - filtered to all methane}
\item{ghg_quantity}{emissions of methane in metric tons of methane per year}
\item{year}{Year of data}
}
}
\source{
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/c_subpart_level_information}
}
\usage{
GHGRP_combustion
}
\description{
A subset of data from the GHGRP
}
\details{
Data for facilities with stationary combustion
}
\keyword{datasets}
