% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGRP_LDC}
\alias{GHGRP_LDC}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Reporting Program (GHGRP)}
\format{
\subsection{\code{GHGRP_LDC} A data frame with 2,222 rows, and 7 columns:}{

\describe{
\item{facility_id}{Unique GHGRP ID for each utility}
\item{reporting_year}{Year of data}
\item{Miles_of_Mains}{miles of natural gas mains pipelines}
\item{N_of_above_grade_T_D_transfer_stations}{Count of above grade (meaning above the finished ground level of structures) transfer and distribution stations that switch gas from transmission lines to distribution lines}
\item{N_of_above_grade_non_T_D_MR_stations}{Count of above grade (meaning above the finished ground level of structures) metering and regulating stations that measure gas flow and regulate the gas pressure in the line}
\item{N_of_below_grade_T_D_transfer_stations}{Count of below grade (meaning below the finished ground level of structures) transfer and distribution stations that switch gas from transmission lines to distribution lines}
\item{N_of_below_grade_non_T_D_MR_stations}{Count of below grade (meaning below the finished ground level of structures) metering and regulating stations that measure gas flow and regulate the gas pressure in the line}
}
}
}
\source{
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/w_local_dist_companies_details}
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/ef_w_equip_leaks_ngdist_leaks}
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/ef_w_equip_leaks_pop_count}
}
\usage{
GHGRP_LDC
}
\description{
A subset of data from the GHGRP
}
\details{
Data for natural gas utilities
}
\keyword{datasets}
