% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HIFLD_NG_data}
\alias{HIFLD_NG_data}
\title{Data formerly provided publicly by the Homeland Infrastructure
Foundation-Level Data (HIFLD) that has since been discontinued.  For this
work it has been processed to map to GHGRP facilities}
\format{
\code{HIFLD_NG_data} A data frame with 2,302 rows, and 5 columns:
\describe{
\item{LATITUDE}{Facility location coordinates}
\item{LONGITUDE}{Facility location coordinates}
\item{GHGRP ID}{Unique GHGRP ID for each facility if there is a corresponding facility}
\item{STATE}{State that the facility operates in}
\item{NAME}{Facility name}
}
}
\source{
\url{https://www.dhs.gov/gmo/hifld}
}
\usage{
HIFLD_NG_data
}
\description{
A subset of data from HIFLD
}
\details{
Data for transmission compressor stations
}
\keyword{datasets}
