% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LMOP_data}
\alias{LMOP_data}
\title{Data from the Environmental Protection Agency's (EPA) Landfill Methane
Outreach Program (LMOP)}
\format{
\subsection{\code{LMOP_data} A data frame with 2,323 rows, and 5 columns:}{

\describe{
\item{GHGRP ID}{Unique GHGRP ID for each facility if there is a corresponding facility}
\item{Latitude}{Facility location coordinates}
\item{Longitude}{Facility location coordinates}
\item{Landfill Name}{Facility name}
\item{Year Landfill Opened}{Year the landfill opened}
}
}
}
\source{
\url{https://www.epa.gov/lmop/landfill-technical-data}
}
\usage{
LMOP_data
}
\description{
A subset of data from LMOP
}
\details{
Data for landfills, including those too small to report to the GHGRP
}
\keyword{datasets}
