% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_individual_sectors.R
\name{log_plot}
\alias{log_plot}
\title{Create a log10 visual}
\usage{
log_plot(
  input,
  title,
  zlim_min = NULL,
  zlim_max = NULL,
  filename,
  plot_directory,
  domain,
  County_Tigerlines,
  State_CB
)
}
\arguments{
\item{input}{SpatRaster.  Intended to be a SpatRaster of gridded methane
emissions.}

\item{title}{\strong{Optional} character providing the main title for the
plot.  The plot will have no main title if not provided.}

\item{zlim_min}{\strong{Optional} numeric providing the minimum value for the
colorscale of the plot, representing the log-scale gridded methane in
nmol/m2/s.  Will default to the minimum of the data if not provided.}

\item{zlim_max}{\strong{Optional} numeric equivalent to zlim_min, but for the
maximum value.}

\item{filename}{\strong{Optional} character providing the output filename for
the plot.  Will default to a png with the name of the input if not provided,
unless the plot is for an entire sector, rather than subsector.  This is
determined by checking input's name includes "Summed". This will use the
same default filename, but place it in a separate subfolder.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{County_Tigerlines}{SpatVector.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization.}

\item{State_CB}{SpatVector.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization.}
}
\value{
This function returns nothing, but does produce a plot with the
gridded methane data colored on a log scale with NA values set to black. All
axes are clearly labeled and state, county, and, if relevant, focus city
boundaries are overlaid in greys and white.
}
\description{
This is a simple helper function for ease of use.  It runs
prep_plot_data and builds useful map visuals of the input data.
}
\details{
This function is intended to be used when verbose = TRUE for
CH4_inventory_build.  It allows for reasonable default scales/titles or set
ones, e.g., if wanting to hold the colorscale constant across multiple
similar sectors/subsectors.
}
\seealso{
\code{\link[=prep_plot_data]{prep_plot_data()}} calculate the log base 10 and remove infinities (i.e.,
0's).

\code{\link[=not_log_plot]{not_log_plot()}} plots the data in simple linear-scale visuals.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
