% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMMT_config.R
\name{M3T_get_config}
\alias{M3T_get_config}
\title{Get package options}
\usage{
M3T_get_config(option = NULL)
}
\arguments{
\item{option}{Character vector of option names to retrieve. If \code{NULL}, returns all options.}
}
\value{
A named list of option values or a single option value if one name is provided.
}
\description{
Retrieve one or more options stored in the package environment.
}
\examples{
M3T_get_config("Process_landfills")
M3T_get_config()
}
