% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3T_config.R
\name{M3T_set_config}
\alias{M3T_set_config}
\title{Set package options}
\usage{
M3T_set_config(...)
}
\arguments{
\item{...}{Named option values to set.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Set one or more options in the package environment.
}
\examples{
M3T_set_config(Process_landfills = FALSE)
M3T_set_config(Source_GHGRP_facility_data = "default", Source_GHGI = "default")
}
