% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landfill_emissions_r1.R
\name{Municipal_solid_waste}
\alias{Municipal_solid_waste}
\title{Create gridded municipal solid waste methane emissions maps}
\usage{
Municipal_solid_waste(
  input_directory,
  domain,
  domain_template,
  state_name_list,
  output_directory,
  inventory_year,
  GHGI_data_yr,
  verbose,
  GHGI_landfill_total,
  GHGRP_facility_data,
  GHGRP_combustion_emissions,
  Source_GHGRP_landfills,
  Source_LMOP,
  landfill_ghgrp_reported,
  landfill_ghgrp_modeled,
  landfill_ghgrp_collection_efficiency,
  plot_directory,
  County_Tigerlines,
  State_CB
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{GHGI_data_yr}{Integer providing the year of data to use for the
\href{https://www.epa.gov/ghgreporting}{Environmental Protection Agency
(EPA) Greenhouse Gas Reporting Program (GHGRP) data}.}

\item{verbose}{Logical indicating whether to save visuals. This includes up to
5 plots of the gridded methane emissions on log scales, one for LMOP
facilities, one for each variation of the GHGRP facilities, and one with the
sum of LMOP and GHGRP emissions. The GHGRP data is the first available in
the order of reported, modeled, then collection efficiency.}

\item{GHGI_landfill_total}{Numeric.  Pulled from \code{\link{M3T_config}}.}

\item{GHGRP_facility_data}{Data.frame with the GHGRP location data for all
years and states as prepared in \code{\link{CH4_inventory_build}} using the
\code{Source_GHGRP_facility_data} provided in \code{\link{M3T_config}}.}

\item{GHGRP_combustion_emissions}{Data.frame with the GHGRP combustion
emissions data for all years and states as prepared in
\code{\link{CH4_inventory_build}} using
\code{Source_GHGRP_combustion_emissions} provided in
\code{\link{M3T_config}}.}

\item{Source_GHGRP_landfills}{Character.  Pulled from
\code{\link{M3T_config}}.}

\item{Source_LMOP}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{landfill_ghgrp_reported}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{landfill_ghgrp_modeled}{Logical.  \code{\link{M3T_config}}.}

\item{landfill_ghgrp_collection_efficiency}{Logical. \code{\link{M3T_config}}.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
4 main netcdf files and 3 sector total netcdf files of gridded methane
emissions from municipal landfills. The main files are titled
"MSW_GHGRP_reported.nc", "MSW_GHGRP_modeled.nc",
"MSW_GHGRP_collection_efficiency.nc", and "MSW_LMOP.nc". The sector total
netcdf files are titled "Landfill_sector_total_GHGRP_reported.nc",
"Landfill_sector_total_GHGRP_modeled.nc", and
"Landfill_sector_total_GHGRP_collection_efficiency.nc".
}
\description{
\code{Municipal_solid_waste} is an internal function that we
strongly recommend users do not use directly, instead using
\code{\link{CH4_inventory_build}} and \code{\link{M3T_config}} which call
this function. \code{Municipal_solid_waste} writes up to 4 main netcdf files
and 3 sector total netcdf files of gridded methane emissions from municipal
landfills, as well as optional visuals
}
\details{
This function calculates and grids methane emissions from municipal
landfills. It uses the \href{https://www.epa.gov/ghgreporting}{Environmental
Protection Agency's (EPA) Greenhouse Gas Reporting Program (GHGRP)}
emissions when available.  It then calculates the difference between
national GHGRP emissions and the
\href{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}{EPA
national Greenhouse Gas Inventory (GHGI)} emissions (GHGI - GHGRP) and
distributes this residual equally to all facilities in the
\href{https://www.epa.gov/lmop/landfill-technical-data}{EPA Landfill Methane
Outreach Program (LMOP)} that are not already accounted for by the GHGRP.

Landfills have 2 options for reporting their emissions to the GHGRP -
equation HH-6 and HH-8.  HH-6 is based on a first order decay model, HH-8 is
based on the collection efficiency of a gas collection system. Landfills can
choose to treat either as the reported value, but both are provided to the
GHGRP.  The modeled values (HH-6) or collection efficiency (HH-8) can be
used rather than the reported values when a gas collection system exists.
Note the GHGRP total used for the LMOP calculation will be the reported
values regardless, as these are what are used in the GHGI.

The GHGRP includes only facilities that emit at least 25,000 metric tons of
carbon dioxide equivalent while the GHGI is intended to capture all national
emissions and LMOP is a voluntary program with location and other details,
but no emissions information.  GHGRP data is available starting in 2011 and
generally is about 2 years behind present day, the GHGI is available
starting in 1990 and is updated approximately in sync with the GHGRP, and
the LMOP is updated more frequently, sometimes multiple times per year.  All
three datasets are annual.  The GHGRP and LMOP are at the facility scale
while the GHGI is national totals.
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.
}
\keyword{internal}
