% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEI_all_years}
\alias{NEI_all_years}
\title{Data from the Environmental Protection Agency's (EPA) National Emissions
Inventory (NEI)}
\format{
\subsection{\code{NEI_all_years} A data frame with 2,323 rows, and 5 columns:}{

\describe{
\item{COUNTY FIPS}{Federal Information Processing Standards (FIPS) code for the county}
\item{EMISSIONS}{Emissions of carbon monoxide (CO) in short tons per year}
\item{INVENTORY YEAR}{Year of data}
\item{SECTOR}{Economic sector and fuel combination}
\item{STATE}{State abbreviation}
\item{STATE FIPS}{Federal Information Processing Standards (FIPS) code for the state}
}
}
}
\source{
\url{https://www.epa.gov/air-emissions-inventories/national-emissions-inventory-nei}
}
\usage{
NEI_all_years
}
\description{
A subset of data from the NEI
}
\details{
Data on CO emissions from a wide variety of sources separated by fuel,
county, year, and economic sector
}
\keyword{datasets}
