% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neighboring_states}
\alias{Neighboring_states}
\title{States that neighbor each state}
\format{
\code{Neighboring_states} A data frame with 49 rows and 49 columns.  Each
row and column is a state abbreviation, a true means the states neighbor
one another (e.g., WA neighbors only ID and OR).
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
Neighboring_states
}
\description{
A logical matrix of which states are neighbors within CONUS
}
\details{
Processed data from the Census tigerlines
}
\keyword{datasets}
