% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NG_distribution_emissions_r4.R
\name{NG_distribution}
\alias{NG_distribution}
\title{Create gridded natural gas distribution methane emissions maps}
\usage{
NG_distribution(
  domain,
  domain_template,
  state_name_list,
  input_directory,
  output_directory,
  inventory_year,
  verbose,
  GHGRP_facility_data,
  GHGRP_subpartW_emissions,
  NG_annex_file,
  Source_EIA_NG_file,
  Source_PHMSA_file,
  Source_GHGRP_LDC,
  GHGI_MnR,
  GHGI_maintenance,
  GHGI_meters,
  GHGI_services,
  State_Tigerlines,
  NG_distribution_by_LDC,
  NG_distribution_by_state,
  NG_distribution_by_domain,
  natural_gas_pipeline_emission_factors,
  natural_gas_res_post_meter_emission_factor,
  natural_gas_com_post_meter_emission_factor,
  Use_ACES,
  Use_Vulcan,
  aces_res,
  aces_com,
  vu_res,
  vu_com,
  plot_directory,
  County_Tigerlines,
  State_CB
)
}
\arguments{
\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{state_name_list}{Character vector listing all states within the desired
domain}

\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in 'CH4_inventory_build'.}

\item{output_directory}{Character providing the full filepath to save
processed data}

\item{inventory_year}{Numeric indicating the desired year of data to use.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions for each
fuel-sector-inventory-variation combination as well as 2 summed plots for
each inventory-variation combination - one for wood and one for all other
sectors.}

\item{GHGRP_facility_data}{Data.frame with the GHGRP location data for all
years and states.  See
\url{https://www.epa.gov/enviro/envirofacts-data-service-api}}

\item{Source_EIA_NG_file}{Character providing the full filepath to the EIA Form 176 -
Annual Report of Natural and supplemental Gas Supply and Disposition xlsx
file available at
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}.
The year can be set to just the desired year.  Then all data can be
downloaded to an xlsx file using a button to the topright of the data table.
If running at the LDC level, the file must be edited such that there is a
consistent identifier with other input data (HIFLD, PHMSA, GHGRP).  There is
an example file in the package's datasets folder that has been successfully
used in this code available for reference.}

\item{Source_PHMSA_file}{Character providing the full filepath to the PHMSA Gas
Distribution Annual Data xlsx file available at
\url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}.
Zip files for groups of years are available via links at the bottom of the
page.  If running at the LDC level the file must be edited such that there
is a consistent identifier with other input data (EIA, HIFLD, GHGRP).  There
is an example file in the package's datasets folder that has been
successfully used in this code available for reference.}

\item{GHGI_MnR}{Character or data.frame.  Pulled from config file. Either GHGI
to indicate the GHGI file should be used to pull emissions and activity data
or a data frame providing the needed values.}

\item{GHGI_maintenance}{Character or data.frame.  Pulled from config file.
Either GHGI to indicate the GHGI file should be used to pull emissions and
activity data or a data frame providing the needed values.}

\item{GHGI_meters}{Character or data.frame.  Pulled from config file. Either
GHGI to indicate the GHGI file should be used to pull emissions and activity
data or a data frame providing the needed values.}

\item{GHGI_services}{Character or data.frame.  Pulled from config file. Either
GHGI to indicate the GHGI file should be used to pull emissions and activity
data or a data frame providing the needed values.}

\item{State_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.}

\item{NG_distribution_by_LDC}{Logical.  Pulled from config file.  Indicating
whether emissions should be calculated at the Local Distribution Company
level.}

\item{NG_distribution_by_state}{Logical.  Pulled from config file.  Indicating
whether emissions should be calculated at the state level.}

\item{NG_distribution_by_domain}{Logical.  Pulled from config file. Indicating
whether emissions should be calculated at the domain level.}

\item{natural_gas_pipeline_emission_factors}{Data frame.  Pulled from config
file.  2 columns by 4 rows.  Columns are "Leaks_per_mile" and
"Avg_emissions_mol_per_s".  Rownames are "Bare_Steel", "Cast_Iron",
"Coated_steel", and "Plastic".  Default is from Weller et al.}

\item{Use_ACES}{Logical indicating whether or not to use ACES to disaggregate
from county-level to pixel level emissions.  Either ACES or Vulcan must be
used, though both can be.}

\item{Use_Vulcan}{Logical indicating whether or not to use Vulcan to
disaggregate from county-level to pixel level emissions.  Either ACES or
Vulcan must be used, though both can be.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}. United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}. US Census Cartographic Boundary files for
visualization
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}.
Only relevant if verbose=T}

\item{natural_gas_post_meter_emission_factor}{Numeric.  Pulled from config
file.  Emission factor for whole-house residential emissions in mol/s.
Default is from Fischer et al.}

\item{ACES_directory}{\strong{Optional} character providing the full path to a
folder containing the ACES sectoral CO2 inventories.  Must include the
residential and commercial sectors.  ACES v2.0 is available at
\url{https://doi.org/10.3334/ORNLDAAC/1943}, though the hourly file should
be averaged across hours to create an annually averaged inventory. Code to
do this on a linux-based HPC system is available as the script
"Annualize_ACES_seawulf.R" and the accompanying batch script
"Annualize_ACES.sh".  The year closest to "inventory_year" is used, but
those further from that year are considered if the closest is unavailable.
Only needed if Use_ACES = TRUE.  At least one of Use_Vulcan or Use_ACES must
be TRUE.}

\item{vulcan_directory}{\strong{Optional} character providing the full path to
a folder containing the Vulcan sectoral CO2 inventories.  Must include the
residential and commercial sectors. Vulcan v3.0 is available at
\url{https://doi.org/10.3334/ORNLDAAC/1741}, and the annual mean files
should be used.  The year closest to "inventory_year" is used.  As all years
are contained in the same file, it does not search for other years.  Only
needed if Use_Vulcan = TRUE.  At least one of Use_Vulcan or Use_ACES must be
TRUE.}

\item{ACES_year}{\strong{Optional} numeric providing the year of ACES data to
use. Only needed if Use_ACES = TRUE.  At least one of Use_Vulcan or Use_ACES
must be TRUE.}

\item{vulcan_year}{\strong{Optional} numeric providing the band of Vulcan data
to use (1-6 = 2010 - 2015).  Only needed if Use_Vulcan = TRUE.  At least one
of Use_Vulcan or Use_ACES must be TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
63 netcdf files of the methane emissions from natural gas distribution. They
are titled as "NG_dist_type_sector_variation_inventory.nc" where type is
upset, serv (services), post_meter, MnR (metering and regulating stations),
and mains; sector is abbreviated as res (residential) or com (commercial);
variation is byLDC, bystate, or bydomain; and inventory is ACES or Vulcan.
}
\description{
\code{NG_distribution} writes up to 63 netcdf files of gridded methane
emissions from natural gas distribution sources, as well as optional visuals
}
\details{
This function calculates and grids methane emissions from natural gas
distribution.  It uses a Homeland Infrastructure Foundation-Level Data
(HIFLD) dataset titled Natural Gas Local Distribution Company Service
Territories, Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI), the EPA Greenhouse Gas Reporting Program (GHGRP), the
Pipeline and Hazardous Materials Safety Administration (PHMSA) gas
distribution annual report, the Energy Information Administration (EIA) Form
176 - Annual Report of Natural and Supplemental Gas Supply and Disposition,
and either the Vulcan or Anthropogenic Carbon Emission System (ACES) CO2
inventory.  This can be done at the Local Distribution Company (LDC) level,
State level, or Domain level.  As several of the input datasets have no
common LDC identifier, they must be carefully matched to run this function
at the LDC level.  If running at the state or domain level, data can be
aggregated and applied between datasets without any manual matching needed.

First all of the input data excluding the GHGI are loaded in and filtered to
only the states within the domain.  The GHGRP data is used to calculate the
typical number of stations per mile of pipeline for each LDC.  Stations
include transmission and distribution transfer stations and metering and
regulating stations.  This is calculated for all stations combined, as well
as separately for above grade and below grade stations.  If calculating by
LDC, the PHMSA miles of pipeline per LDC is used in the calculation of
stations per mile, rather than the GHGRP, though the miles of pipeline for
each LDC in GHGRP and PHMSA are compared and if any LDC differs by > 5\\% an
error will be flagged.

The GHGI Annex data is then pulled and includes emission factors and
activity data for metering an regulating stations separated by inlet
pressure, and just the emission factors for pipeline services separated by
pipeline material, meters separated by customer type (residential,
commercial, industrial), and different maintenance events (pressure relief
valve release, blow-downs, accidental dig-ins).

PHMSA data on miles of pipeline separated by material is then combined with
the input "natural_gas_pipeline_emission_factors", which by default are the
emission factors from Table 2 of Weller et al. who used research equipped
Google Street View cars to measure > 4000 leaks in the U.S.  PHMSA data on
pipeline services, separated by pipeline material, are combined with the
GHGI emission factors to get emissions for each LDC.

GHGRP data is aggregated to the state level to get the average number of
above or below grade stations per mile.  If not calculating emission by LDC,
this is combined with the PHMSA miles of pipeline to get an estimate of the
number of above and below grade stations for each LDC in the PHMSA dataset.

The combined miles of pipeline including services is calculated from the
PHMSA and the various input datasets are subset to only the necessary
variables, aggregated to the state level and merged.  The variables kept
differ slightly if calculating by LDC.

If not calculating by LDC, then there is no HIFLD csv file necessary,
leaving the PHMSA activity data, the EIA sales data, and the GHGRP activity
data.

If calculating by LDC a HIFLD csv and shapefile (containing the same
information) are also included.  The residuals (LDCs that could not be
matched across datasets) are assigned to an "OTHER" LDC that includes all
land in the state not already accounted for by an LDC.  The number of
stations for each LDC is now overwritten to be the values for that specific
LDC from the GHGRP.  State average stations per mile (calculated from the
GHGRP) are then applied only to LDCs that did not report to the GHGRP.

Service and pipeline emissions are then split into residential and
commercial fractions.  This is calculated for residential as the sum of all
emissions * N residential customers / N total residential and commercial
customers.  The calculation would be equivalent for commercial customers.

Emissions for metering and regulating stations are then calculated as a
function of pressure.  The number of stations of a grade (above or below)
are multiplied with the national fraction of metering and regulating
stations of that grade that are a certain pressure window and then
multiplied by the emission factor for that pressure window.  E.g., N Above
grade (by LDC or state) * national type fraction * national type emission
factor.  These emissions are then split into residential and commercial in
the same manner as service and pipeline emissions were.

Meter emissions are calculated separately for each customer type
(residential, commercial, industrial) using the corresponding GHGI emission
factors.  These emissions are then split into residential and commercial,
incorporating industrial emissions into residential and commercial
proportionally.  This was done as the industrial sectors of ACES/Vulcan are
dominated by point sources, many of which don't even rely on natural gas. It
was chosen to split by emissions rather than number of customers as this was
considered more representative (e.g., there may be many residential
customers, yet the commercial ones consume more natural gas overall).

Maintenance and upsets are calculated using national GHGI emission factors
and split into residential and commercial in the same manner as service and
pipeline emissions were.

Post-meter emissions are calculated, strictly for residential, using the
volume of gas delivered to residential customers and the
natural_gas_post_meter_emission_factor, which is based on Fischer et al. by
default.  Fischer et al. measured whole-house emissions from 75 homes in
California using mass balance.

Finally, ACES and/or Vulcan residential and commercial gridded CO2 emission
maps are loaded in.  The emissions for each subsector are then distributed
using the ACES or Vulcan CO2 inventory.  This can be done at the LDC, state,
or domain level.  Emissions at this point are at the state level if not
calculating by LDC and can then be aggregated to the domain level. Otherwise
emissions are at the LDC level and can be aggregated to the state or domain
level.  As such, producing output at the state and LDC level will result in
slightly different output than running only at the state level.

So, to summarize this relatively complex sector, PHMSA data on miles of
pipeline by type and number of services by type of pipeline is combined with
emission factors to calculate emissions.  GHGRP state average numbers of M&R
facilities per mile are also combined with an emission factor, broken down
by pressure.  If calculating byLDC and GHGRP data exists for an LDC, the
reported counts are used instead.  These emissions are then distributed to
the appropriate LDC territories using HIFLD shapefiles, or aggregated at the
state level. To further disaggregate the emissions, they are broken into
residential and commercial portions using the fraction of customers as
reported to the EIA. They are then distributed to the pixel scale using the
residential and commercial sectors of the CO2 inventories Vulcan/ACES.  As
such, GHGRP emissions are not directly used at all, PHMSA is the source of
most activity data with GHGRP providing the numbers of M&R facilities, HIFLD
solely provides shapefiles for each LDC if operating by LDC, and EIA
provides a breakdown of residential vs commercial customers.

GHGRP data is available for this sector starting in 2011 and generally is
about 2 years behind present day, the GHGI is available starting in 1990 and
is updated approximately in sync with the GHGRP.  The HIFLD dataset is
updated infrequently and is available for 2019 and 2017.  The EIA form is
annually reported and is available starting in 1997 and is generally updated
in September to the previous year (i.e., Sept 2024 adds 2023 data).  The
PHMSA data is annual and is available starting in 1970 and is available up
to the most recent year.  The GHGRP includes only facilities that emit at
least 25,000 metric tons of carbon dioxide equivalent while the GHGI is
intended to capture all national emissions. All other datasets are meant to
be inclusive of national facilities. All data is annual.  The GHGI is
national totals while all other datasets are at the facility scale.

GHGRP data, GHGI data, and the HIFLD shapefile will be automatically
downloaded.

The GHGRP is available at \url{https://ghgdata.epa.gov/ghgp/main.do}. For
individual LDC metering and regulating station counts, among other
variables, one must filter to the natural gas local distribution companies
sector, select an individual facility and select "View reported data".  The
GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}.
The necessary GHGI Annex data is available at
\url{https://www.epa.gov/ghgemissions/natural-gas-and-petroleum-systems-ghg-inventory-additional-information-1990-2022-ghg}
for the 2024 GHGI.  In the GHGI Annexes, available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022},
for 2024 there is a link to the file in Section 3.6: "Methodology for
Estimating CH4, CO2, and N2O Emissions from Natural Gas Systems".  The excel
file has multiple sheets, each of which has a separate layout.  The HIFLD
dataset is available at
\url{https://hifld-geoplatform.hub.arcgis.com/datasets/geoplatform::natural-gas-service-territories/about}
and can be donwloaded as both a shapefile or a csv. EIA form 176 is
available at
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}
and can be downloaded as an excel file.  The PHMSA Gas Distribution Annual
Data can be download at
\url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
as a zip file with an excel file for each year. ACES is available at
\url{https://doi.org/10.3334/ORNLDAAC/1943} and Vulcan is available at
\url{https://doi.org/10.3334/ORNLDAAC/1741}.

See references \href{https://doi.org/10.1021/acs.est.0c00437}{Weller et
al.}, \href{https://doi.org/10.1021/acs.est.8b03217}{Fischer et al.},
\href{https://doi.org/10.1029/2020JD032974}{Vulcan} and,
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\examples{
library(terra)
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)

GHGI_EF_dataframe <- data.frame("Leaks_per_mile"=
                                  c(0.51,1,0.61,0.43),
                                "Avg_emissions_mol_per_s"=
                                  c(2.24,1.72,2,2.03)/(16.043*60)) #converting from g/min to mol/s
rownames(GHGI_EF_dataframe) <- c("Bare_Steel",
                                 "Cast_Iron",
                                 "Coated_steel",
                                 "Plastic")

NG_distribution(domain=grid_vect,
                domain_template=grid,
                state_name_list=c("DE","MD","NJ","NY","PA"),
                input_directory="~/../Desktop/in/",
                output_directory="~/../Desktop/out/",
                inventory_year=2018,
                verbose=TRUE,
                GHGRP_facility_data="~/../Desktop/in/GHGRP/facility_info.csv",
                EIA_file = "~/../Desktop/in/176 Type of Operations and Sector Items.xlsx",
                PHMSA_file = "~/../Desktop/in/annual_gas_distribution_2010_present/annual_gas_distribution_2019.xlsx",
                GHGI_MnR="GHGI",
                GHGI_maintenance="GHGI",
                GHGI_meters="GHGI",
                GHGI_services="GHGI",
                State_Tigerlines=vect("~/../Desktop/in/State_Tigerlines/tl_2018_us_state.shp"),
                NG_distribution_by_LDC = FALSE,
                NG_distribution_by_state = TRUE,
                NG_distribution_by_domain = TRUE,
                natural_gas_pipeline_emission_factors=GHGI_EF_dataframe,
                natural_gas_post_meter_emission_factor=7850/401*0.005/(16.043*60*60*24*365),
                Use_ACES=TRUE,
                Use_Vulcan=TRUE,
                ACES_directory="~/../Desktop/in/Inventories/ACES_v2.0",
                vulcan_directory="~/../Desktop/in/Inventories/Vulcan_v3.0",
                ACES_year=2017,
                vulcan_year=6,
                County_Tigerlines=vect("~/../Desktop/in/County_Tigerlines/tl_2018_us_county.shp"),
                plot_directory="~/../Desktop/plots/")
}
\references{
\href{https://doi.org/10.1021/acs.est.0c00437}{Weller et al.}

\href{https://doi.org/10.1021/acs.est.8b03217}{Fischer et al.}

\href{https://doi.org/10.1029/2020JD032974}{Vulcan}

\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\seealso{
\itemize{
\item \code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
\item \code{\link[=disaggregation]{disaggregation()}} Disaggregates data using NEI county level CO data.
}
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
