% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD_fractions_by_state.R
\name{NLCD_fractions_by_state}
\alias{NLCD_fractions_by_state}
\title{Process the National Land Cover Database for \code{Wastewater} function}
\usage{
NLCD_fractions_by_state(
  input_directory,
  Source_wastewater_NLCD,
  domain,
  domain_template,
  State_Tigerlines,
  state_name_list,
  output_directory
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{Source_wastewater_NLCD}{Character.  Pulled from
\code{\link{M3T_config}}.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}
}
\value{
Nothing is returned from the function, but the main outputs are 1
netcdf file per state in the domain and a csv.  The netcdf files are titled
as "X_NLCD_suburban.nc" where X is the state abbreviation (e.g., MD, DE).
Each file contains a SpatRaster with the fractional coverage of the
developed open space + developed low intensity for the state on the same
grid as the input domain. The csv is titled "NLCD_state_total_areas.csv" and
provides the total area of each land cover type in each state.
}
\description{
\code{NLCD_fractions_by_state} writes 1 netcdf file per state within the
domain as well as a csv.  These netcdf files have the 30 m National Land
Cover Database Developed, Open Space and Developed, Low Intensity land cover
types for each state.  The csv provides the total area covered by these two
land cover types for each state separately.  This information is used in
\code{Wastewater} to map septic emissions.
}
\details{
This function first crops the 30 m
\href{https://www.sciencebase.gov/catalog/item/655ceb8ad34ee4b6e05cc51a}{National
Land Cover Database} to the states within the domain.  Then the data is
subset to create a SpatRaster that is 1 for Developed, Open Space or
Developed, Low Intensity land cover and 0 for all other land cover types.
This is then split into separate SpatRasters for each state and the total
area of these two land cover types for each state is calculated.  The data
is then reprojected to the domain's projection. Finally these SpatRasters
are saved for use in the \code{\link{Wastewater}} function. The total area
of the two land cover types for each state, including data outside the
domain if relevant, is saved as a csv.
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.

\code{\link[=Wastewater]{Wastewater()}} Calculates methane emissions for the wastewater sector.
}
\keyword{internal}
