% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_individual_sectors.R
\name{not_log_plot}
\alias{not_log_plot}
\title{Create a visual}
\usage{
not_log_plot(
  input,
  title,
  zlim_min = NULL,
  zlim_max = NULL,
  filename,
  plot_directory,
  domain,
  County_Tigerlines,
  State_CB,
  Urban_Tigerlines
)
}
\arguments{
\item{input}{SpatRaster.  Intended to be a SpatRaster of gridded methane
emissions.}

\item{title}{\strong{Optional} character providing the main title for the
plot.  The plot will have no main title if not provided.}

\item{zlim_min}{\strong{Optional} numeric providing the minimum value for the
colorscale of the plot, representing the gridded methane in nmol/m2/s.  Will
default to the minimum of the data if not provided.}

\item{zlim_max}{\strong{Optional} numeric equivalent to zlim_min, but for the
maximum value.}

\item{filename}{\strong{Optional} character providing the output filename for
the plot.  Will default to a png with the name of the input if not provided,
unless the plot is for an entire sector, rather than subsector.  This will
use the same default filename, but place it in a separate subfolder.}
}
\value{
This function returns nothing, but does produce a plot with the
gridded methane data colored with NA values set to black. All axes are
clearly labeled and state, county, and, if relevant, focus city boundaries
are overlaid in greys and white.
}
\description{
This is a simple helper function for ease of use.  It builds
useful map visuals of the input data.
}
\details{
This function is intended to be used when verbose = TRUE for
CH4_inventory_build.  It allows for reasonable default scales/titles or set
ones, e.g., if wanting to hold the colorscale constant across multiple
similar sectors/subsectors.
}
\examples{
not_log_plot(septic_flux,filename="Wastewater_dom_septic_national",
             "Domestic Wastewater - Septic\n national EPA septic distributed using \ndeveloped open space/low intensity land cover",
             global(min(septic_flux,septic_flux2),min),
             global(max(septic_flux,septic_flux2),max))
}
\seealso{
\code{\link[=log_plot]{log_plot()}} plots the data in simple log-scale visuals.
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
