% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetland_fraction_r2_WIP.R
\name{NWI_Wetland_fraction}
\alias{NWI_Wetland_fraction}
\title{Create gridded wetland coverage maps using National Wetlands Inventory
data}
\usage{
NWI_Wetland_fraction(
  input_directory,
  output_directory,
  domain,
  domain_template,
  state_name_list,
  Use_SOCCR1,
  Use_SOCCR2
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data}

\item{output_directory}{Character providing the full filepath to save
processed data}

\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{state_name_list}{Character vector listing all states within the desired
domain}

\item{Use_SOCCR1}{Logical.  Pulled from config file.  Indicating whether or
not to calculate emissions using SOCCR1.}

\item{Use_SOCCR2}{Logical.  Pulled from config file.  Indicating whether or
not to calculate emissions using SOCCR2.}

\item{Include_freshwater}{Logical.  Pulled from config file.  Indicating
whether or not to calculate emissions for freshwater wetlands using
Rosentreter et al.}
}
\value{
Nothing is returned from the function, but the main outputs are TIFF
files of the fractional wetland coverage per pixel for each wetland type and
state.  They are titled "state abbreviation _ wetland type
abbreviation.tiff".
}
\description{
\code{NWI_Wetland_fraction} writes multiple tiff files of the
fractional wetland cover in each pixel.  There is one file for each state -
wetland type combination.
}
\details{
This function downloads National Wetlands Inventory (NWI) data and
calculates the fraction of each gridcell in the domain that is covered by
the NWI polygons.  This is done separately for each state as the data is
available by state or watershed.  It also processes each type of wetland
separately.

The wetland types are pulled from the NWI system, subsystem, and class as
defined here
\url{https://www.fws.gov/media/national-wetland-inventory-wetlands-and-deepwater-map-code-diagram}
and discussed in detail here
\url{https://www.fws.gov/media/classification-wetlands-and-deepwater-habitats-united-states}.
They are M1 (marine, subtidal), M2 (marine, intertidal), E1 (estuarine,
subtidal), E2 (estuarine, intertidal), R1 (riverine, tidal), R2 (riverine,
lower perrennial), R3 (riverine, upper perennial), R4 (riverine,
intermittent), L1 (lacustrine, limnetic), L2 (lacustrine, littoral), PFO
(palustrine, forested), and PNF (palustrine, all non-forested classes).

The NWI data for each state will be automatically downloaded.  This can be
up to several GB per state.

Given the high resolution polygons and significant number being processed,
this function can be time consuming.

The NWI is a biannually updated dataset including the extent and type of
wetlands across the United States.  It is available at
\url{https://www.fws.gov/program/national-wetlands-inventory/wetlands-data.}

A separate TIFF file is saved for each state - wetland type combination.
These are used in further processing.
}
\examples{
library(terra)
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)

NWI_Wetland_fraction(input_directory="~/../Desktop/in/",
                     output_directory="~/../Desktop/out/",
                     Use_SOCCR1=TRUE,
                     Use_SOCCR2=TRUE,
                     Include_freshwater=TRUE,
                     domain=grid_vect,
                     domain_template=grid,
                     state_name_list=c("DE","MD","NJ","NY","PA"))
}
\seealso{
\itemize{
\item \code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
\item \code{\link[=SOCCR_Wetlands]{SOCCR_Wetlands()}} Calculates methane emissions for the wetland sector using the state of the carbon cycle report.
}
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
