% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHMSA_natural_gas_distribution}
\alias{PHMSA_natural_gas_distribution}
\title{Pipelines and Hazardous Materials Safety Administration (PHMSA) Gas
Distribution Annual Data}
\format{
\subsection{\code{PHMSA_natural_gas_distribution} A data frame with 21,321 rows and}{

14 columns:
\describe{
\item{OPERATOR_NAME}{Utility name}
\item{REPORT_YEAR}{Year the data was reported for}
\item{NUM_SRVCS_TOTAL}{Total number of service lines (those connecting from a distribution main to an end user) in the system at the end of the year}
\item{AVERAGE_LENGTH}{Average length of service lines in feet}
\item{STOP}{State in which the system operates}
\item{MMILES_TOTAL}{Total miles of pipe in the system at the end of the year}
\item{MMILES_bare_steel}{Total miles of bare steel (cathodically protected or not) and copper pipe in the system at the end of the year}
\item{MMILES_iron}{Total miles of cast iron, ductile iron, and reconditioned cast iron pipe in the system at the end of the year}
\item{MMILES_coat_steel}{Total miles of coated steel (cathodically protected or not) and "other" pipe in the system at the end of the year}
\item{MMILES_plastic}{Total miles of plastic pipe in the system at the end of the year}
\item{NUM_SRVS_unp_steel}{Total number of not cathodically protected steel (bare or coated) service lines in the system at the end of the year}
\item{NUM_SRVS_cp_steel}{Total number of cathodically protected steel (bare or coated) and "other" service lines in the system at the end of the year}
\item{NUM_SRVS_plastic}{Total number of plastic service lines in the system at the end of the year}
\item{NUM_SRVS_copper_iron}{Total number of copper, cast iron, ductile iron, and reconditioned cast iron service lines in the system at the end of the year}
}
}
}
\source{
\url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
}
\usage{
PHMSA_natural_gas_distribution
}
\description{
A subset of tables from the PHMSA reports
}
\details{
The PHMSA reports on natural gas distribution service and main pipelines by
utility
}
\keyword{datasets}
