% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_GEPA.R
\name{Prepare_GEPA}
\alias{Prepare_GEPA}
\title{Pull gridded methane emissions maps for sectors not built with other
functions from the Maasakkers gridded EPA inventory}
\usage{
Prepare_GEPA(
  inventory_year,
  input_directory,
  output_directory,
  plot_directory,
  Source_GEPA,
  County_Tigerlines,
  State_CB,
  domain,
  domain_template,
  verbose
)
}
\arguments{
\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{Source_GEPA}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{verbose}{Logical indicating whether to save visuals.  They include 2
plots of the gridded methane emissions, 1 for the GEPA sectors used in M3T,
1 for all GEPA sectors on the same colorscale as the other.}
}
\value{
Nothing is returned from the function, but the main outputs are 3
netcdf files of the methane emissions from the gridded EPA product.  They
are titled "GEPA_thermo.nc" for gridded EPA thermogenic,
"GEPA_non_thermo.nc" for non-thermogenic, "GEPA_ind_landfill.nc" for
industrial landfills.
}
\description{
\code{Prepare_GEPA} writes 3 netcdf files of gridded methane emissions
\itemize{
\item 1 for thermogenic sectors, 1 for industrial landfills, and 1 for
non-thermogenic sources.
}
}
\details{
This function downloads the most appropriate year of gridded methane
data at 0.1 deg by 0.1 deg from Maasakkers et al. and pulls sectors that
have not been calculated in other scripts.  They are projected to the domain
grid and combined into thermogenic and non-thermogenic groups, with the
exception of industrial landfills which are saved separately.  For
thermogenic this includes
\itemize{
\item mobile combustion
\item Coal (abandoned, surface, and underground)
\item Petroleum Systems (exploration, production, refining, and transport)
\item Abandoned Oil and Gas
\item Natural Gas (exploration, processing, and production)
\item Industry (petrochemical, and ferroalloy)
}
And for non-thermogenic this includes
\itemize{
\item Composting
\item Enteric Fermentation
\item Manure Management
\item Rice Cultivation
\item Field Burning
}

The data is available at \url{https://doi.org/10.5281/zenodo.8367082}.  The
closest year will be automatically downloaded.

See reference \href{https://doi.org/10.1021/acs.est.3c05138}{Maasakkers et
al.}
}
\references{
\href{https://doi.org/10.1021/acs.est.3c05138}{Maasakkers et al.}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.
}
\keyword{internal}
