% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_config.R
\name{Set_config}
\alias{Set_config}
\title{Create the config file utilized by \code{CH4_inventory_build}}
\usage{
Set_config(run_directory)
}
\arguments{
\item{run_directory}{Character providing the full filepath to load/save
input and output data.  Subfolders will be created.}
}
\value{
A config file is created and opened for editing in the user set
directory and that directory is saved to the global environment.  These
user-edited settings are utilized throughout.
}
\description{
This function is a wrapper to build the config that provides
\code{CH4_inventory_build} with all emission factors and similar user-edited
input data as well as options for which sectors to run and, when applicable,
which variations to run.
}
\details{
This function creates the config, which is intended to be the
location for all user edited variables other than inputs to the main
\code{CH4_inventory_build} function.
}
\examples{
Set_config("~/../Desktop/CONUS_2020_MMMT_run/")

}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
