% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Total_national_open_or_low_int_area}
\alias{Total_national_open_or_low_int_area}
\title{The total "Developed, Open" and "Developed, Low Intensity" in the Continental
US (CONUS) land area}
\format{
\subsection{\code{Total_national_open_or_low_int_area} A list with 14 entries named}{

by year, each with numeric values representing the total of this land cover
for the year across CONUS.
}
}
\source{
\url{https://doi.org/10.5066/P94UXNTS}
}
\usage{
Total_national_open_or_low_int_area
}
\description{
National Land Cover Database (NLCD) calculated land cover data
}
\details{
Annual NLCD data is processed to calculate the total fraction of land cover
that is the land types considered likely to have septic systems rather than
sewer systems.
}
\keyword{datasets}
