% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NG_transmission_emissions_r1.R
\name{Transmission}
\alias{Transmission}
\title{Create gridded natural gas transmission methane emissions maps}
\usage{
Transmission(
  input_directory,
  GHGI_transmission_compressors,
  GHGI_Pipeline,
  Source_HIFLD_compressor_file,
  Source_EIA_transmission_file,
  domain,
  domain_template,
  GHGRP_facility_data,
  GHGRP_subpartW_emissions,
  GHGRP_combustion_emissions,
  state_name_list,
  output_directory,
  inventory_year,
  GHGI_data_yr,
  verbose,
  plot_directory,
  County_Tigerlines,
  State_CB
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{GHGI_transmission_compressors}{Character or data.frame.  Pulled from
config file. Either GHGI to indicate the GHGI file should be used to pull
emissions and activity data or a data frame providing the needed values.}

\item{GHGI_Pipeline}{Character or data.frame.  Pulled from
config file. Either GHGI to indicate the GHGI file should be used to pull
emissions and activity data or a data frame providing the needed values.}

\item{Source_HIFLD_compressor_file}{Character providing the full filepath to the
HIFLD compressor data.  As this file is now deprecated and no replacement
has been created, it currently must be provided as part of the package.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{GHGRP_facility_data}{Data.frame with the GHGRP location data for all
years and states.  See
https://www.epa.gov/enviro/envirofacts-data-service-api}

\item{GHGRP_combustion_emissions}{Data.frame with the GHGRP combustion
emissions data for all years and states as prepared in
\code{\link{CH4_inventory_build}} using
\code{Source_GHGRP_combustion_emissions} provided in
\code{\link{M3T_config}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{GHGI_data_yr}{Integer providing the year of data to use for the
\href{https://www.epa.gov/ghgreporting}{Environmental Protection Agency
(EPA) Greenhouse Gas Reporting Program (GHGRP) data}.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions for each
fuel-sector-inventory-variation combination as well as 2 summed plots for
each inventory-variation combination - one for wood and one for all other
sectors.}

\item{plot_directory}{Character providing the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{County_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile downloaded in Main.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are 2
netcdf files of the methane emissions from natural gas transmission.  They
are titled "NG_trans_compressors.nc" and "NG_trans_pipes.nc" where
NG=natural gas and trans=transmission.   The first file is for transmission
compressor emissions and the second is for transmission pipeline emissions.

2 csv files are also optionally saved.  These are "NG_trans_compressors.csv"
and "NG_trans_compressors_all.csv".  The simpler csv includes only the name,
location, and assigned emissions for compressors within the domain that were
pulled from the GHGRP  The _all files include all variables that were in the
corresponding input file for the same compressors.
}
\description{
\code{Transmission} writes 2 netcdf files of gridded methane emissions
from natural gas transmission, as well as optional visuals and 2 optional
csv files.
}
\details{
This function calculates and grids methane emissions from natural gas
transmission systems. It uses the Homeland Infrastructure Foundation-Level
Data (HIFLD), Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI), EPA Greenhouse Gas Reporting Program (GHGRP) and Energy
Information Administration (EIA) Energy Atlas data.

The necessary GHGRP, GHGI, and EIA data will be automatically downloaded.

For pipelines the GHGI data includes the emissions and activity data for
pipeline leaks, meter and regulating stations, and venting.  An emission
factor in mols of methane per meter of pipeline per second can then be
calculated.  This emission factor is then applied to the EIA
inter/intrastate pipeline location data.

For compressors the GHGI data includes the emissions and activity data for
compressor station fugitive emissions, dehydrator vents, flaring, engines,
turbines, generators, pneumatic devices, and venting.  For generators they
are split into engine and turbine emissions so the ratio between
transmission/storage emissions for engines and turbines are applied to the
generator values to get the transmission component of generator emissions.
An national average emission rate in mols of methane per station per second
can then be calculated.  This emission rate is then assigned to all HIFLD
compressors.  GHGRP data is then used to overwrite this default emission
rate.  Note most compressor stations do not report their emissions to the
GHGRP, so only a subset will be overwritten using GHGRP data.  The GHGRP
compressor emissions are scaled so that the average emissions within the
domain are equal to the national average calculated from the GHGI.  As there
is not a common identifier between the GHGRP and HIFLD datasets, the nearest
facility is considered the matching facility.

The GHGRP includes only facilities that emit at least 25,000 metric tons of
carbon dioxide equivalent while the GHGI is intended to capture all national
emissions.  GHGRP data is available starting in 2010 and generally is about
2 years behind present day, the GHGI is available starting in 1990 and is
updated approximately in sync with the GHGRP.  Both datasets are annual. The
GHGRP is at the facility scale while the GHGI is national totals.

The HIFLD compressor data being used was last updated December 2022.  It has
since been deprecated, but as there is no replacement available, it is still
being used.

The GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}.
The necessary GHGI Annex data is available at
\url{https://www.epa.gov/ghgemissions/natural-gas-and-petroleum-systems-ghg-inventory-additional-information-1990-2022-ghg}
for the 2024 GHGI.  In the GHGI Annexes, available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022},
for 2024 there is a link to the file in Section 3.6: "Methodology for
Estimating CH4, CO2, and N2O Emissions from Natural Gas Systems".  The excel
file has multiple sheets, each of which has a separate layout.  The GHGRP is
available at \url{https://ghgdata.epa.gov/ghgp/main.do}, and the EIA data is
available at
\url{https://atlas.eia.gov/datasets/eia::natural-gas-interstate-and-intrastate-pipelines/about}.
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.
}
\keyword{internal}
