% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wastewater.R
\name{Wastewater}
\alias{Wastewater}
\title{Create gridded methane emissions maps for wastewater treatment plants
and septic systems}
\usage{
Wastewater(
  input_directory,
  output_directory,
  Wastewater_use_CWNS,
  Wastewater_use_DMR,
  Wastewater_Municipal_Method_Moore,
  Wastewater_Municipal_Method_GHGI,
  Wastewater_national_septic,
  Wastewater_state_septic,
  domain,
  domain_template,
  GHGRP_facility_data,
  Source_GHGRP_wastewater,
  Source_CWNS,
  Source_DMR,
  Source_wastewater_NLCD,
  Source_State_population_data,
  inventory_year,
  National_wastewater_info,
  Wastewater_reported_State_info,
  GHGI_wastewater_data,
  GHGI_data_yr,
  Total_national_open_or_low_int_area,
  State_Tigerlines,
  state_name_list,
  County_Tigerlines,
  plot_directory,
  State_CB,
  verbose
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{Wastewater_use_CWNS}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Wastewater_use_DMR}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Wastewater_Municipal_Method_Moore}{Logical.  Pulled from
\code{\link{M3T_config}}.}

\item{Wastewater_Municipal_Method_GHGI}{Logical.  Pulled from
\code{\link{M3T_config}}.}

\item{Wastewater_national_septic}{Logical.  Pulled from
\code{\link{M3T_config}}.}

\item{Wastewater_state_septic}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{GHGRP_facility_data}{Data.frame with the GHGRP location data for all
years and states as prepared in \code{\link{CH4_inventory_build}} using the
\code{Source_GHGRP_facility_data} provided in \code{\link{M3T_config}}.}

\item{Source_GHGRP_wastewater}{Character.  Pulled from
\code{\link{M3T_config}}.}

\item{Source_CWNS}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_DMR}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_State_population_data}{Character.  Pulled from
\code{\link{M3T_config}}.}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{National_wastewater_info}{Data.frame.  Pulled from
\code{\link{M3T_config}}.}

\item{Wastewater_reported_State_info}{Data.frame.  Pulled from
\code{\link{M3T_config}}.}

\item{GHGI_wastewater_data}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{GHGI_data_yr}{Integer providing the year of data to use for the
\href{https://www.epa.gov/ghgreporting}{Environmental Protection Agency
(EPA) Greenhouse Gas Reporting Program (GHGRP) data}.}

\item{Total_national_open_or_low_int_area}{Numeric.  Pulled from
\code{\link{M3T_config}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes up to 5 plots of the gridded methane emissions.  Up to 2 for
municipal wastewater treatment facilities, up to 2 for septic, and 1 for
industrial wastewater treatment facilities}
}
\value{
Nothing is returned from the function, but the main outputs are up to
14 netcdf files of the methane emissions from wastewater.  They are titled
"Wastewater_ind.nc" for industrial wastewater,
"Wastewater_X_Y_dom_central.nc" for municipal wastewater,
"Wastewater_dom_septic_Z.nc" for septic emissions.  X is the input data used
for the wastewater treatment plants - CWNS or DMR, Y is the method of
converting to emissions - GHGI for downscaling GHGI totals proportionally
with flow or Moore for applying the moore et al. log-linear relationship
between flow and emissions, and Z is the approach used for septic emissions
\itemize{
\item bystate or national.
The 8 possible combinations are named similarly as
"Wastewater_sector_total_X_Y_Z.nc".
}
}
\description{
\code{Wastewater} is an internal function that we strongly
recommend users do not use directly, instead using
\code{\link{CH4_inventory_build}} and \code{\link{M3T_config}} which call
this function. \code{Wastewater} writes up to 15 netcdf files of gridded
methane emissions - 1 - 4  for municipal wastewater treatment facilities, 1
for industrial wastewater treatment facilities, 1 - 2 for septic systems,
and then up to 8 possible combinations across these.
}
\details{
This function calculates and grids methane emissions from wastewater
treatment and septic systems. It takes the total flow through municipal
facilities to either distribute national total emissions proportionally or
to calculate emissions using a log-log linear relationship from the
scientific literature.  For industrial facilities, the Environmental
Protection Agency's (EPA) Greenhouse Gas Reporting Program (GHGRP) data is
used. Lastly septic emissions are mapped to "developed, open space" and
"developed, low intensity" land cover.  The emissions are either
disaggregated from national septic emissions or estimated using
state-specific values as available.

For municipal wastewater treatment facilities there are a few variations.
Input data can be set to either the Environmental Protection Agency's (EPA)
Clean Watershed Needs Survey (CWNS), available at
\url{https://www.epa.gov/cwns/clean-watersheds-needs-survey-cwns-past-reports-and-data}
or the National Pollutant Discharge Elimination System (NPDES) Discharge
Monitoring Reports (DMR) available through the Enforcement and Compliance
History Online (ECHO) tool at
\url{https://echo.epa.gov/trends/loading-tool/water-pollution-search}.  The
CWNS is produced infrequently while DMR data is regularly updated, though
the facilities and their data can differ between these products.  Facility
locations and total effluent flow is used from these files.  The calculation
method can then either be set to disaggregate the EPA Greenhouse Gas
Inventory (GHGI) national total emissions to these facilities, using flow as
a proxy, or calculate emissions for each facility using a log-log linear
relationship between flow and emissions (log10(CH4 emissions in g/s) =
1.28*log10(flow in m3/s) + 0.93) based on Moore et al.  Moore et al.
measured emission rates at 96 wastewater treatment facilities to develop
this relationship, which suggests nearly 2x the emissions as the GHGI using
DMR flow rates as of 2023.  Note the authors' first paper discussed this
relatinoship and was updated using the complete dataset from a followup
paper.

GHGRP data is used for any and all industrial wastewater treatment
facilities.

For septic emissions the output from \code{\link{NLCD_fractions_by_state}}
is used. This provides the fractional coverage of the 2 landcover types
(developed, open space and developed, low intensity) for each state in the
domain as well as gridded data for the whole domain. Their are 2 approaches
to calculate methane emissions from these.

National option - The total landcover of these types across the continental
US is calculated using the NLCD combined with estimates using the limited AK
NLCD data available.  We use this national total land area in km2 and the
GHGI national total septic emissions in mol/s to calculate an emission
factor in mol/s/km2.  This emission factor can be combined with the gridded
land cover data to get gridded septic emissions.  Note this ignores any of
this land cover in other US territory outside of the continental US like
Guam, Hawaii, or American Somoa.

State option - we can instead take state-reported data rather than using
national  Some states report the fraction of people in the state using
septic systems to the American Housing Survey in the Plumbing, Water, and
Sewage Disposal survey, available here
\url{https://www.census.gov/programs-surveys/ahs/data/interactive/ahstablecreator.html?s_areas=00000&s_year=2021&s_tablename=TABLE1&s_bygroup1=1&s_bygroup2=1&s_filtergroup1=1&s_filtergroup2=1}.
Note this data is reported every other year, not annually. For states that
don't, the 1990 census data on septic fraction is available here
\url{https://www.census.gov/data/tables/time-series/dec/coh-sewage.html} and
this will be scaled using the change in the national septic fraction from
1990 (from census) to the desired year (from the housing survey).  This new
fraction can be combined with up to date population data available from the
census
\url{https://www.census.gov/data/tables/time-series/demo/popest/2020s-state-total.html},
the total amount of this land cover in the state, and the GHGI septic EF (g
CH4 per person per day) to get emissions.

The GHGRP includes only facilities that emit at least 25,000 metric tons of
carbon dioxide equivalent while the GHGI is intended to capture all national
emissions.  GHGRP data is available starting in 2010 and generally is about
2 years behind present day, the GHGI is available starting in 1990 and is
updated approximately in sync with the GHGRP.  Census data comes out
annually and is generally 1 year behind.  All three datasets are annual. The
GHGRP is at the facility scale while the GHGI is national totals.  Census
data can be obtained at many different scales, but is used here at the state
and national scale.

The GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}
The GHGRP is available at \url{https://www.epa.gov/ghgreporting} The Annual
NLCD is available at \doi{doi:10.5066/P94UXNTS} The 2011 AK NLCD
is available at \doi{doi:10.5066/P97S2IID} The 2016 AK NLCD is
available at \doi{doi:10.5066/P96HHBIE}

See references Homer et al. at \doi{doi:10.1016/j.isprsjprs.2020.02.019} and
Moore et al. at \doi{doi:10.1021/acs.est.2c05373}.
}
\references{
Homer et al.; \doi{doi:10.1016/j.isprsjprs.2020.02.019}

Moore et al.; \doi{doi:10.1021/acs.est.2c05373}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\code{\link[=NLCD_fractions_by_state]{NLCD_fractions_by_state()}} Calculates the fraction of low intensity urban
land cover in each pixel.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.
}
\keyword{internal}
