% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wastewater_1990_state_septic}
\alias{Wastewater_1990_state_septic}
\title{Historical state septic data from the U.S. Census Bureau}
\format{
\code{Wastewater_1990_state_septic} A data frame with 52 rows and 2
columns:
\describe{
\item{State}{State name}
\item{Septic_Fraction}{Fraction of the population using septic systems}
}
}
\source{
\url{https://www.census.gov/data/tables/time-series/dec/coh-sewage.html}
}
\usage{
Wastewater_1990_state_septic
}
\description{
A subset of data from the Census Bureau
}
\details{
The estimated fraction of the population using septic systems at the state
level for 1990 published by the U.S. Census Bureau
}
\keyword{datasets}
