% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastewater_state_septic_area}
\alias{wastewater_state_septic_area}
\title{The total "Developed, Open" and "Developed, Low Intensity" land cover by
state in the Continental US}
\format{
\subsection{\code{wastewater_state_septic_area} A list with 14 entries named by}{

year, each with a vector representing the total of this land cover for each
state. Rownames are state abbreviations.
}
}
\source{
\url{https://doi.org/10.5066/P94UXNTS}
}
\usage{
wastewater_state_septic_area
}
\description{
National Land Cover Database (NLCD) calculated land cover data
}
\details{
Annual NLCD data is processed to calculate the total fraction of land cover
that is the land types considered likely to have septic systems rather than
sewer systems separately for each state.
}
\keyword{datasets}
