% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH4_inventory_build.R
\name{CH4_inventory_build}
\alias{CH4_inventory_build}
\title{Main function to calculate gridded methane emissions.  Runs sector by
sector according to the config file.}
\usage{
CH4_inventory_build(
  run_directory,
  inventory_year,
  domain,
  domain_res,
  domain_crs = "epsg:4326",
  verbose = FALSE
)
}
\arguments{
\item{run_directory}{Character providing the full filepath to load/save input
and output data.  Subfolders will be created.}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{domain}{data.frame or character.  If data.frame, provides the corner
coordinates of the desired region to process.  The first column would be x
values, the second column y values, and the first row would be the minima,
the second row the maxima.  These must be provided in the appropriate units
for the domain_crs parameter (e.g., decimal degrees for a lat/long
projection).

If character, it can be defined based on
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{U.S.
Census Bureau Tigerlines}.  It can be a state abbreviation, state name,
state FIPS code (as a character), Urban Area name, or Urban Area Census
Code.  For example: "DE", "Delaware", and "10" are all equivalent and "Long
Neck, DE" and "51202" are equivalent. It can also be "CONUS" to run for the
entire continental United States or "custom" for a domain created by the
user interactively. Lastly, it can be a filepath pointing to a polygon file
that can be interpreted by the terra package. Names must match exactly and
FIPS or urban area codes must include leading 0's.  Lists of the names and
codes for all urban areas for the most recent census are available
\href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural.html}{here}.
Links to previous census urban areas are on the same page.  A list with
state fips codes is available
\href{https://www.census.gov/library/reference/code-lists/ansi.html}{here}.}

\item{domain_res}{Numeric providing the resolution for the domain.  Can be
length 1 for equal x and y resolution or length 2 (x, y).}

\item{domain_crs}{Character providing the projection of the domain in
\href{https://proj.org/en/stable/operations/projections/index.html}{PROJ
string} or \href{https://epsg.io/}{EPSG codes} or
\href{https://docs.ogc.org/is/12-063r5/12-063r5.html}{WKT} formats. Commonly
desired crs include "epsg:4326" for lat/long and "+proj=lcc +lat_0=40
+lon_0=-97 +lat_1=33 +lat_2=45 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs"
for a m grid in the lambert conformal conical projection used by some CO2
inventories for the continental US}

\item{verbose}{Logical indicating whether to save sector and subsector
visuals.}
}
\value{
Nothing is returned from the function, but there will be frequent user
updates as multiple other functions are run and output files created for the
various sectors.
}
\description{
\code{CH4_inventory_build} runs multiple other functions to calculate
the gridded methane emissions sector-by-sector using information provided in
the config file and inputs.
}
\details{
This function will call multiple internal functions and use
\code{\link{M3T_config}} to create gridded methane inventories. Internet
access is required so that the necessary datasets can be downloaded either
directly from the source or from a companion Zenodo contiaining
pre-processed data unless all "Source_" variables in
\code{\link{M3T_config}} are set to filepaths that point to local copies of
the needed data.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
CH4_inventory_build(run_directory=tempdir(),
                    inventory_year=2019,
                    domain="RI",
                    domain_res=1,
                    domain_crs="epsg:4326",
                    verbose=F)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{M3T_config} Generates the config function with user-editable
settings used throughout processing.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
