% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WETCHARTS_downscaling.R
\name{Disaggregate_Wetcharts}
\alias{Disaggregate_Wetcharts}
\title{Create model-averaged downscaled Wetcharts wetland methane maps}
\usage{
Disaggregate_Wetcharts(
  input_directory,
  output_directory,
  domain,
  domain_template,
  verbose,
  plot_directory,
  inventory_year,
  County_Tigerlines,
  State_CB,
  Use_NLCD,
  Source_wetland_NLCD,
  Wetcharts_model_subset
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data}

\item{output_directory}{Character providing the full filepath to save
processed data}

\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{verbose}{Logical indicating whether to save additional output.  This
includes monthly plots of the gridded methane emissions on log scales, saved
separately for each model subset and each land cover dataset used.}

\item{plot_directory}{Character providing the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{inventory_year}{Character indicating the desired year of data to use.}

\item{County_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{State_CB}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{Use_NLCD}{Logical. Pulled from config file. Indicating whether to use
the National Land Cover Database (NLCD) to downscale Wetcharts.}

\item{Wetcharts_model_subset}{Numeric list. Pulled from config file. Indicates
which models of Wetcharts to average across.  Multiple list entries are
allowed, providing multiple variations of wetcharts to be run
simultaneously.  This is far faster than running them separately as
landcover processing is more time consuming and avoided if running multiple
wetcharts subsets simultaneously.}

\item{Use_NALCMS}{Logical. Pulled from config file. Indicating whether to use
the North America Land Change Monitoring System (NALCMS) to downscale
Wetcharts.}

\item{NLCD_file}{Character.  The NLCD land cover data as an img. Available at
\url{https://www.mrlc.gov/data?f%5B0%5D=category%3ALand%20Cover&f%5B1%5D=region%3Aconus
 }.}

\item{NALCMS_file}{Character.  The NALCMS land cover data as a tif.  Available
at \url{http://www.cec.org/north-american-land-change-monitoring-system/}.}

\item{Wetcharts_file}{Character providing the full filepath to the Wetcharts
model file}
}
\value{
Nothing is returned from the function, but the main outputs are
netcdf files of the methane emissions from wetlands with 1 file per land
cover used and per model subset.  They are titled
"Wetcharts_landcover_Downscaled_subset_A.nc" where landcover is NLCD or
NALCMS and A is just a numeric to identify which model subset, in case
multiple were set.

If verbose is set to TRUE, then multiple figures are also saved.  Log scale
plots with consistent axes are saved for each model subset and land cover
used.  They are named "Wetcharts_landcover_subset_A_annual.png" where
landcover is NLCD or NALCMS, A is a numeric to identify the model subset.
}
\description{
\code{Disaggregate_Wetcharts} writes 1 monthly netcdf file of gridded
wetland methane emissions per model subset set in the config and per land
classification dataset used for the downscaling.  Includes optional visuals
as well.
}
\details{
This function takes the Wetcharts model ensemble and land cover data
to downscale the wetland emissions from 0.5 deg x 0.5 deg to 0.1 deg x 0.1
deg.  The data is also cropped to the domain, converted to the proper units,
and averaged across wetcharts models using the user-defined subset(s).  The
National Land Cover Database (NLCD) and/or the North American Land Change
Monitoring System (NALCMS) can be used for the land cover data.

The 30 m x 30 m land classification data is aggregated to 0.1 deg x 0.1 deg
and used to calculate the fraction of each 0.5 deg x 0.5 deg pixel's total
wetland area that is within each 0.1 deg x 0.1 deg pixel.  Any 0.5 deg x 0.5
deg pixel without land cover data is distributed equally to the 0.1 deg x
0.1 deg pixels.

Wetcharts data is then subset to the user-defined subset(s) models and
averaged.  This is multiplied by the wetland fraction calculated using the
land cover data to disaggregate fluxes to 0.1 deg x 0.1 deg and saved.  It
is first converted from flux to emissions, then downscaled, then converted
back to fluxes as fluxes are per area and this must be considered when
changing the area of pixels. There is a separate file for each model subset
and each land cover dataset used.

The appropriate year of Wetcharts v1.3.1 will be automatically downloaded.

Wetcharts is available at \url{https://doi.org/10.3334/ORNLDAAC/2346}, the
NLCD is available at
\url{https://www.mrlc.gov/data?f\%5B0\%5D=category\%3ALand\%20Cover&f\%5B1\%5D=region\%3Aconus},
and the NALCMS is available at
\url{http://www.cec.org/north-american-land-change-monitoring-system/}.
}
\examples{
library(terra)
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)

Disaggregate_Wetcharts(input_directory="~/../Desktop/in/",
                       output_directory="~/../Desktop/out/",
                       domain=grid_vect,
                       domain_template=grid,
                       verbose=TRUE,
                       inventory_year=2018,
                       plot_directory="~/../Desktop/plots/",
                       State_CB=vect("~/../Desktop/in/State_CB/tl_2018_us_state.shp"),
                       County_Tigerlines=vect("~/../Desktop/in/County_Tigerlines/tl_2018_us_county.shp"),
                       Use_NLCD=TRUE,
                       Use_NALCMS=TRUE,
                       NLCD_file=file.path("~/../Desktop/in/nlcd_2019_land_cover_l48_20210604/nlcd_2019_land_cover_l48_20210604.img"),
                       NALCMS_file=file.path("~/../Desktop/in/NALCMS_2020_land_cover/NA_NALCMS_landcover_2020_30m/data/NA_NALCMS_landcover_2020_30m.tif"),
                       Wetcharts_model_subset=list(c(1913,1914,1923,1924,1933,1934,2913,2914,2923,
                                                     2924,2933,2934,3913,3914,3923,3924,3933,3934)))


}
\seealso{
\itemize{
\item \code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
\item \code{\link[=SOCCR_Wetlands]{SOCCR_Wetlands()}} Calculates methane emissions for the wetland sector using the state of the carbon cycle report instead.
}
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
