% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disaggregate_Wetcharts.R
\name{Disaggregate_Wetcharts}
\alias{Disaggregate_Wetcharts}
\title{Create model-averaged downscaled Wetcharts wetland methane maps}
\usage{
Disaggregate_Wetcharts(
  input_directory,
  output_directory,
  domain,
  domain_template,
  verbose,
  plot_directory,
  inventory_year,
  County_Tigerlines,
  State_Tigerlines,
  State_CB,
  Source_wetland_NLCD,
  Source_wetcharts,
  Wetcharts_model_subset
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes monthly plots of the gridded methane emissions on log scales, saved
separately for each model subset used.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}

\item{Source_wetland_NLCD}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_wetcharts}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Wetcharts_model_subset}{List of numeric vectors.  Pulled from
\code{\link{M3T_config}}.}
}
\value{
Nothing is returned from the function, but the main outputs are
netcdf files of the methane emissions from wetlands with 1 file per land
cover used and per model subset.  They are titled
"Wetcharts_NLCD_Downscaled_subset_#.nc" where # is just a numeric to
identify which model subset, in case multiple were set.

If verbose is set to TRUE, then multiple figures are also saved.  Log scale
plots with consistent axes are saved for each model subset used.  They are
named "Wetcharts_NLCD_subset_#_annual.png" where # is a numeric to identify
the model subset.
}
\description{
\code{Disaggregate_Wetcharts} writes 1 monthly netcdf file of gridded
wetland methane emissions per model subset set in the config and per land
classification dataset used for the downscaling.  Includes optional visuals
as well.
}
\details{
This function takes the Wetcharts model ensemble and land cover data
to downscale the wetland emissions from 0.5 deg x 0.5 deg to 0.1 deg x 0.1
deg.  The data is also cropped to the domain, converted to the proper units,
and averaged across wetcharts models using the user-defined subset(s).  The
National Land Cover Database (NLCD) and/or the North American Land Change
Monitoring System (NALCMS) can be used for the land cover data.

The 30 m x 30 m land classification data is aggregated to 0.1 deg x 0.1 deg
and used to calculate the fraction of each 0.5 deg x 0.5 deg pixel's total
wetland area that is within each 0.1 deg x 0.1 deg pixel.  Any 0.5 deg x 0.5
deg pixel without land cover data is distributed equally to the 0.1 deg x
0.1 deg pixels.

Wetcharts data is then subset to the user-defined subset(s) models and
averaged.  This is multiplied by the wetland fraction calculated using the
land cover data to disaggregate fluxes to 0.1 deg x 0.1 deg and saved.  It
is first converted from flux to emissions, then downscaled, then converted
back to fluxes as fluxes are per area and this must be considered when
changing the area of pixels. There is a separate file for each model subset
and each land cover dataset used.

The appropriate year of Wetcharts v1.3.1 will be automatically downloaded.

Wetcharts is available at \url{https://doi.org/10.3334/ORNLDAAC/2346}, the
NLCD is available at
\url{https://www.mrlc.gov/data?f\%5B0\%5D=category\%3ALand\%20Cover&f\%5B1\%5D=region\%3Aconus},
and the NALCMS is available at
\url{http://www.cec.org/north-american-land-change-monitoring-system/}.
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.

\code{\link[=SOCCR_Wetlands]{SOCCR_Wetlands()}} Calculates methane emissions for the wetland sector using
the state of the carbon cycle report instead.
}
\keyword{internal}
