% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHGRP_landfills}
\alias{GHGRP_landfills}
\title{Data from the Environmental Protection Agency's (EPA) Greenhouse Gas
Reporting Program (GHGRP)}
\format{
\code{GHGRP_landfills} A data frame with 16,444 rows, and 7 columns:
\describe{
\item{facility_id}{Unique GHGRP ID for each facility}
\item{year}{Year of data}
\item{facility_name}{Name of facility}
\item{ghg_name}{greenhouse gas name - filtered to all methane}
\item{ghg_quantity}{emissions of methane in metric tons of methane per year}
\item{HH_modeled}{emissions of methane in metric tons of methane per year as estimated using a first order decay model (method HH-6), required for all reporting landfills}
\item{HH_collection_efficiency}{emissions of methane in metric tons of methane per year as estimated using a back calculation with the amount of landfill gas captured and an estimated collection efficiency (method HH-8), required for all reporting landfills with a gas collection system}
}
}
\source{
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/hh_subpart_level_information}
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/hh_gas_collection_system_detls}
}
\usage{
GHGRP_landfills
}
\description{
A subset of data from the GHGRP
}
\details{
Data for landfills
}
\keyword{datasets}
