% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_based_disaggregation.R
\name{Inventory_based_disaggregation}
\alias{Inventory_based_disaggregation}
\title{Disaggregate emissions to the pixel scale using ACES or Vulcan CO2
inventories}
\usage{
Inventory_based_disaggregation(
  input_inventory,
  totals,
  agg_level,
  NEI_input,
  cover_all,
  out_envir
)
}
\arguments{
\item{input_inventory}{SpatRaster.  Either a Vulcan or ACES sectoral
inventory.  ACES is available at \url{https://doi.org/10.3334/ORNLDAAC/1943}
and Vulcan is available at \url{https://doi.org/10.3334/ORNLDAAC/1741}.
Annual mean files should be used for Vulcan.  The hourly ACES file should be
averaged across hours to create an annual mean inventory. There is a
function to do this as well, though it is quite time consuming given the
amount of data that needs to be downloaded.}

\item{totals}{Character vector.  Various subsectors to run through the
disaggregation process.}

\item{agg_level}{Character.  The scale of the data before disaggregation,
solely used for naming.}

\item{NEI_input}{SpatVector.  Includes attributes for each spatial feature
that detail the emissions at the current agg_level separated by totals.}

\item{cover_all}{List of matrices.  Each entry in the list is a separate
state/county and the matrices are the output of the function
terra::extract(weights=T,exact=T,cells=T).  These provide the fraction of
each pixel contained within the state/county.  This allows for better
accounting for pixels that are on the borders of states/counties.}

\item{out_envir}{Environment.  Where to assign output data.  Should be the
environment of \code{stationary_combustion} or \code{Natural_Gas_Distribution} as the output
will be used and exported there.}
}
\value{
Nothing is returned.  The disaggregated methane data is assigned to
the specified out_envir.  This will be a SpatRaster list with one for each
of the totals.
}
\description{
This is a function utilized by \code{stationary_combustion} and
\code{NG_distribution}.  It distributes emissions from a larger scale (state or
county) to the pixel scale using gridded CO2 emissions.
}
\details{
This function is intended to be used by \code{stationary_combustion} and
\code{NG_distribution}.  For each pixel it takes the ratio of CO2 from that pixel
to the total CO2 for the state/county it is in as input.  This is combined
with the state/county total CH4 emissions (also input) to disaggregate the
emissions to the pixel-scale.  If there are no emissions for the
state/county then the entire area is assigned an equal fraction of
emissions.  This is done using sectoral (residential, commercial,
industrial, electric) CO2 inventories.

ACES is available at \url{https://doi.org/10.3334/ORNLDAAC/1943} and Vulcan
is available at \url{https://doi.org/10.3334/ORNLDAAC/1741}.

See references \href{https://doi.org/10.1029/2020JD032974}{Vulcan} and
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\references{
\href{https://doi.org/10.1029/2020JD032974}{Vulcan}

\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\seealso{
\code{\link[=Natural_Gas_Distribution]{Natural_Gas_Distribution()}} Calculates methane emissions for the natural gas
distribution sector.

\code{\link[=Stationary_combustion]{Stationary_combustion()}} Calculates methane emissions for the stationary
combustion sector.
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
\keyword{internal}
