% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3T-package.R
\docType{package}
\name{M3T-package}
\alias{M3T}
\alias{M3T-package}
\title{M3T: Develop Reasonable Methane Inventories for Urban Areas}
\description{
The Modular Methane Mapping Tool (M3T) uses a user defined mix of a variety of literature based methods to estimate gridded methane emissions for several sectors relevant for urban areas. Available only for the Continental United States with user specified resolution (minimum of 0.01 deg or 1 km2), year (> 2011), and domain.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Kristian-hajny/M3T/issues}
}

}
\author{
\strong{Maintainer}: Kristian Hajny \email{kris.hajny@gmail.com} (\href{https://orcid.org/0000-0003-3249-7157}{ORCID})

Other contributors:
\itemize{
  \item Joseph Pitt \email{joseph.pitt@bristol.ac.uk} (\href{https://orcid.org/0000-0002-8660-5136}{ORCID}) [contributor]
  \item The Research Foundation for The State University of New York [copyright holder]
}

}
\keyword{internal}
