% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3T_config_documentation.R
\name{M3T_config}
\alias{M3T_config}
\title{Package Options for M3T}
\description{
M3T stores many user-configurable options in a private
environment. These control which sectors to process, how to access required
datasets, and variations to run. You can get and set these options using the
accessor functions: \code{\link{M3T_get_config}} and
\code{\link{M3T_set_config}}.
}
\details{
The options are described by sector to simplify navigating.  All
relevant datasets have a Source_X option that can generally be set to
"default" to rely on preprocessed data built into the package or its
\href{https://zenodo.org/}{companion Zenodo}, "download" to have the package
download the data from the source, or a filepath pointing to a local copy of
the properly formatted data file.  As datasets can move, change format, etc.
over time, the download option may fail over time.

\bold{Across Sectors}
\itemize{
\item{\bold{Terra_datatype} - character describing how to save raster data.  Default "FLT8S".  See '?terra::terraOptions'.  This is temporarily applied while running 'CH4_inventory_build'.}
\item{\bold{Terra_progress} - integer describing when to use progress bars when processing raster data.  Default 0 (none).  See '?terra::terraOptions'.  This is temporarily applied while running 'CH4_inventory_build'.}
\item{\bold{Base_timeout} - integer describing the maximum time to attempt a download in seconds.  Default 20 minutes. See '?options'.  This is temporarily applied while running 'CH4_inventory_build' and may be critical if any Source_X is set to "download".}

Method Variations
\itemize{
\item{\bold{Use_ACES} - logical stating whether the \href{https://doi.org/10.3334/ORNLDAAC/1943}{ACES CO2 inventory} should be used to downscale the stationary combustion and natural gas distribution emissions.  Default TRUE.}
\item{\bold{Use_Vulcan} - logical stating whether the \href{https://earth.gov/ghgcenter/data-catalog/vulcan-ffco2-yeargrid-v4}{Vulcan CO2 inventory} should be used to downscale the stationary combustion and natural gas distribution emissions.  Default TRUE.}
\item{\bold{Combine_sectors} - logical stating whether sectors should be summed to create all possible unique inventory combinations across variations.  Default TRUE.  Be aware that with many variations this can create >1,000 inventories.}
\item{\bold{Separate_thermo} - logical stating whether combined inventories should also be calculated separately for thermogenic (i.e., natural gas) and non-thermogenic sources.  Default FALSE.  Be aware that this will triple the number of inventories created.}
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - character stating "default", "download", or a vector of filepaths to the needed files.  Includes state, county, and urban tigerlines from the \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{U.S. Census Bureau}. Used throughout.}
\item{\bold{Source_Cartographic_Boundaries_data} - character stating "default", "download", or a filepath pointing to the needed file.  The State Cartographic Boundary file is similar to the state tigerlines, but excludes water areas and is also produced by the \href{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}{U.S. Census Bureau}. Used throughout for visualizations only and used to interactively define the domain if set to "custom".}
\item{\bold{Source_GHGRP_facility_data} - character stating "default", "download", or a filepath pointing to the needed file.  Data on facility locations from the \href{https://enviro.epa.gov/envirofacts/metadata/table/ghg/pub_dim_facility}{EPA Greenhouse Gas Reporting Program's Envirofacts API}. Used for the landfill, natural gas distribution, natural gas transmission, and wastewater sectors as they rely in part on GHGRP data.}
\item{\bold{Source_GHGRP_combustion} - character stating "default", "download", or a filepath pointing to the needed file.  Data on facility combustion emissions (subpart C) from the \href{https://enviro.epa.gov/envirofacts/metadata/table/ghg/c_subpart_level_information}{EPA Greenhouse Gas Reporting Program's Envirofacts API}. Used for the landfill and natural gas transmission sectors as these sources report to subpart C.}
\item{\bold{Source_GHGRP_NG} - character stating "default", "download", or a filepath pointing to the needed file.  Data on petroleum and natural gas system facility-level emissions (subpart W) from the \href{https://enviro.epa.gov/envirofacts/metadata/table/ghg/w_subpart_level_information}{EPA Greenhouse Gas Reporting Program's Envirofacts API}. Used for the natural gas distribution and natural gas transmission sectors as these sources report to subpart W.}
\item{\bold{Source_GHGI} - character stating "default", "download", or a filepath pointing to a directory with the needed files.  The main text and annex tables from the \href{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022}{EPA Greenhouse Gas Inventory}. Optionally used for the landfill, natural gas distribution, natural gas transmission, and stationary combustion sectors to get the national activity and emission factor data used.}
\item{\bold{Source_ACES} - character stating "default", "download", or a filepath pointing to a directory with the needed files.  Annual rasters of the industrial, residential, electric production, and commercial sectors from the \href{https://doi.org/10.3334/ORNLDAAC/1943}{ACES CO2 inventory}.  Optionally used to downscale stationary combustion and natural gas distribution methane emissions.  Code to convert the hourly data to annual is available on the \href{https://zenodo.org/}{companion Zenodo}.}
\item{\bold{Source_Vulcan} - character stating "default", "download", or a filepath pointing to a directory with the needed files.  Annual rasters of the industrial, residential, electric production, and commercial sectors from the \href{https://earth.gov/ghgcenter/data-catalog/vulcan-ffco2-yeargrid-v4}{Vulcan v4.0 CO2 inventory}.  Optionally used to downscale stationary combustion and natural gas distribution methane emissions.}
}
}

\bold{Landfills}
\itemize{
\item{\bold{Process_landfills} - logical stating if this sector should be run.  Default TRUE.}

Method Variations
\itemize{
Landfills have 2 methods to estimate their emissions - a first order decay
model, and an assumed collection efficiency of a gas collection system.
Either can be used as the "reported" value and this can change over time for
a given landfill. You can use the one they report, or force it to one method.
Note landfills without gas collection systems will still be included using
the modeled emission estimate if forcing to the collection efficiency
estimate.
\item{\bold{landfill_ghgrp_reported} - logical stating if the "reported" estimate should be used.}
\item{\bold{landfill_ghgrp_modeled} - logical stating if the decay model estimate should be used.}
\item{\bold{landfill_ghgrp_collection_efficiency} - logical stating if the collection efficiency estimate should be used.}
}

Accessing Datasets
\itemize{
\item{\bold{Source_GHGRP_landfills} - character stating "default", "download", or a filepath pointing to a directory with the needed file. The EPA GHGRP tables \href{https://enviro.epa.gov/envirofacts/metadata/table/ghg/hh_subpart_level_information}{Envirofacts subpart level info} and \href{https://enviro.epa.gov/envirofacts/metadata/table/ghg/hh_gas_collection_system_detls}{Envirofacts gas collection system details} that include GHGRP modeled, collection efficiency, and "reported" emissions.}
\item{\bold{Source_LMOP} - character stating "default", "download", or a filepath pointing to a directory with the needed file.  The \href{https://www.epa.gov/lmop/landfill-technical-data}{EPA Landfill Methane Outreach Program} is a voluntary program with some details including location, but no emissions information.  Used to distribute the residual betwen the GHGI and GHGRP estimates to landfills to small to report to the GHGRP (25,000 MT CO2e).}
}

Emission Factors and similar
\itemize{
\item{\bold{GHGI_landfill_total} - character or numeric listing either "GHGI" to pull the value from the GHGI or an integer value in gigagrams CH4/yr (or kilotons CH4/yr). Represents the net municipal landfill (MSW) emissions that can be pulled from GHGI table "CH4 emissions from Landfills (kt)".}
}
}

\bold{Natural gas distribution - including residential post-meter}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}

\bold{Natural gas transmission}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}

\bold{Stationary combustion}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}

\bold{Wastewater}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}

\bold{Wetlands and inland waters}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}

\bold{Remaining sectors from the gridded EPA inventory}
\itemize{
\item{\bold{Terra_datatype} - }
\item{\bold{Terra_progress} - }
\item{\bold{Base_timeout} - }

Method Variations
\itemize{
\item{\bold{Use_ACES} - .}
\item{\bold{Use_Vulcan} - .}
\item{\bold{Combine_sectors} - }
\item{\bold{Separate_thermo} - }
}

Accessing Datasets
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}

Emission Factors and similar
\itemize{
\item{\bold{Source_Tigerlines_data} - }
\item{\bold{Source_GHGRP_facility_data} - }
\item{\bold{Source_GHGRP_combustion} - }
\item{\bold{Source_GHGI} - }
\item{\bold{Source_GHGRP_NG} - }
\item{\bold{Source_Cartographic_Boundaries_data} - }
\item{\bold{Source_ACES} - }
\item{\bold{Source_Vulcan} - }
}
}
}
\seealso{
\itemize{
\item \link{M3T_get_config} Get config options
\item \link{M3T_set_config} Set config options for the current R session
\item \link{CH4_inventory_build} Calculates methane inventory using settings provided in config.
\item \link[terra:terraOptions]{terra::terraOptions} Options for the terra package
\item \link[base:options]{base::options} Options in base R
}
}
