% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_consistent}
\alias{make_consistent}
\title{Helper function for making GHGRP data consistent across sectors}
\usage{
make_consistent(input)
}
\arguments{
\item{input}{Data.frame. A GHGRP dataframe built from a "subpart information"
table such as
\url{https://enviro.epa.gov/envirofacts/metadata/table/ghg/hh_subpart_level_information}}
}
\value{
input with changes described as in Details
}
\description{
Helper function for making GHGRP data consistent across sectors
}
\details{
This function will change "ghg_gas_name" to "ghg_name" and
"reporting_year" to "year" as these column names are not consistent across
tables.  It will also force the "ghg_name" and "facility_name" data to be
lower case and remove any data for a ghg besides methane.
}
\keyword{internal}
