% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NG_distribution_emissions_r4.R
\name{NG_distribution}
\alias{NG_distribution}
\title{Create gridded natural gas distribution methane emissions maps}
\usage{
NG_distribution(
  domain,
  domain_template,
  state_name_list,
  input_directory,
  output_directory,
  inventory_year,
  GHGI_data_yr,
  verbose,
  GHGRP_facility_data,
  GHGRP_subpartW_emissions,
  Source_EIA_NG_file,
  Source_PHMSA_file,
  Source_GHGRP_LDC,
  GHGI_MnR,
  GHGI_maintenance,
  GHGI_meters,
  GHGI_services,
  State_Tigerlines,
  NG_distribution_by_LDC,
  NG_distribution_by_state,
  NG_distribution_by_domain,
  natural_gas_pipeline_emission_factors,
  natural_gas_res_post_meter_emission_factor,
  natural_gas_com_post_meter_emission_factor,
  Use_ACES,
  Use_Vulcan,
  aces_res,
  aces_com,
  vu_res,
  vu_com,
  plot_directory,
  County_Tigerlines,
  State_CB
)
}
\arguments{
\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{GHGI_data_yr}{Integer providing the year of data to use for the
\href{https://www.epa.gov/ghgreporting}{Environmental Protection Agency
(EPA) Greenhouse Gas Reporting Program (GHGRP) data}.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions for each
fuel-sector-inventory-variation combination as well as 2 summed plots for
each inventory-variation combination - one for wood and one for all other
sectors.}

\item{GHGRP_facility_data}{Data.frame with the GHGRP location data for all
years and states as prepared in \code{\link{CH4_inventory_build}} using the
\code{Source_GHGRP_facility_data} provided in \code{\link{M3T_config}}.}

\item{GHGRP_subpartW_emissions}{Data.frame with the GHGRP petroleum and
natural gas systems emissions data for all years and states as prepared in
\code{\link{CH4_inventory_build}} using
\code{Source_GHGRP_subpartW_emissions} provided in \code{\link{M3T_config}}.}

\item{Source_EIA_NG_file}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_PHMSA_file}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_GHGRP_LDC}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{GHGI_MnR}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{GHGI_maintenance}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{GHGI_meters}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{GHGI_services}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{NG_distribution_by_LDC}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{NG_distribution_by_state}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{NG_distribution_by_domain}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{natural_gas_pipeline_emission_factors}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{natural_gas_res_post_meter_emission_factor}{Numeric.  Pulled from \code{\link{M3T_config}}.}

\item{natural_gas_com_post_meter_emission_factor}{Numeric.  Pulled from \code{\link{M3T_config}}.}

\item{Use_ACES}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Use_Vulcan}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{aces_res}{SpatRaster of residential CO2 emissions from the ACES
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_ACES} and \code{Source_ACES}.}

\item{aces_com}{SpatRaster of commercial CO2 emissions from the ACES
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_ACES} and \code{Source_ACES}.}

\item{vu_res}{SpatRaster of residential CO2 emissions from the Vulcan v4.0
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{vu_com}{SpatRaster of commercial CO2 emissions from the Vulcan v4.0
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
63 netcdf files of the methane emissions from natural gas distribution. They
are titled as "NG_dist_type_sector_variation_inventory.nc" where type is
upset, serv (services), post_meter, MnR (metering and regulating stations),
and mains; sector is abbreviated as res (residential) or com (commercial);
variation is byLDC, bystate, or bydomain; and inventory is ACES or Vulcan.
}
\description{
\code{NG_distribution} writes up to 63 netcdf files of gridded methane
emissions from natural gas distribution sources, as well as optional visuals
}
\details{
This function calculates and grids methane emissions from natural gas
distribution.  It uses a Homeland Infrastructure Foundation-Level Data
(HIFLD) dataset titled Natural Gas Local Distribution Company Service
Territories, Environmental Protection Agency's (EPA) Greenhouse Gas
Inventory (GHGI), the EPA Greenhouse Gas Reporting Program (GHGRP), the
Pipeline and Hazardous Materials Safety Administration (PHMSA) gas
distribution annual report, the Energy Information Administration (EIA) Form
176 - Annual Report of Natural and Supplemental Gas Supply and Disposition,
and either the Vulcan or Anthropogenic Carbon Emission System (ACES) CO2
inventory.  This can be done at the Local Distribution Company (LDC) level,
State level, or Domain level.  As several of the input datasets have no
common LDC identifier, they must be carefully matched to run this function
at the LDC level.  If running at the state or domain level, data can be
aggregated and applied between datasets without any manual matching needed.

First all of the input data excluding the GHGI are loaded in and filtered to
only the states within the domain.  The GHGRP data is used to calculate the
typical number of stations per mile of pipeline for each LDC.  Stations
include transmission and distribution transfer stations and metering and
regulating stations.  This is calculated for all stations combined, as well
as separately for above grade and below grade stations.  If calculating by
LDC, the PHMSA miles of pipeline per LDC is used in the calculation of
stations per mile, rather than the GHGRP, though the miles of pipeline for
each LDC in GHGRP and PHMSA are compared and if any LDC differs by > 5\\% an
error will be flagged.

The GHGI Annex data is then pulled and includes emission factors and
activity data for metering an regulating stations separated by inlet
pressure, and just the emission factors for pipeline services separated by
pipeline material, meters separated by customer type (residential,
commercial, industrial), and different maintenance events (pressure relief
valve release, blow-downs, accidental dig-ins).

PHMSA data on miles of pipeline separated by material is then combined with
the input "natural_gas_pipeline_emission_factors", which by default are the
emission factors from Table 2 of Weller et al. who used research equipped
Google Street View cars to measure > 4000 leaks in the U.S.  PHMSA data on
pipeline services, separated by pipeline material, are combined with the
GHGI emission factors to get emissions for each LDC.

GHGRP data is aggregated to the state level to get the average number of
above or below grade stations per mile.  If not calculating emission by LDC,
this is combined with the PHMSA miles of pipeline to get an estimate of the
number of above and below grade stations for each LDC in the PHMSA dataset.

The combined miles of pipeline including services is calculated from the
PHMSA and the various input datasets are subset to only the necessary
variables, aggregated to the state level and merged.  The variables kept
differ slightly if calculating by LDC.

If not calculating by LDC, then there is no HIFLD csv file necessary,
leaving the PHMSA activity data, the EIA sales data, and the GHGRP activity
data.

If calculating by LDC a HIFLD csv and shapefile (containing the same
information) are also included.  The residuals (LDCs that could not be
matched across datasets) are assigned to an "OTHER" LDC that includes all
land in the state not already accounted for by an LDC.  The number of
stations for each LDC is now overwritten to be the values for that specific
LDC from the GHGRP.  State average stations per mile (calculated from the
GHGRP) are then applied only to LDCs that did not report to the GHGRP.

Service and pipeline emissions are then split into residential and
commercial fractions.  This is calculated for residential as the sum of all
emissions * N residential customers / N total residential and commercial
customers.  The calculation would be equivalent for commercial customers.

Emissions for metering and regulating stations are then calculated as a
function of pressure.  The number of stations of a grade (above or below)
are multiplied with the national fraction of metering and regulating
stations of that grade that are a certain pressure window and then
multiplied by the emission factor for that pressure window.  E.g., N Above
grade (by LDC or state) * national type fraction * national type emission
factor.  These emissions are then split into residential and commercial in
the same manner as service and pipeline emissions were.

Meter emissions are calculated separately for each customer type
(residential, commercial, industrial) using the corresponding GHGI emission
factors.  These emissions are then split into residential and commercial,
incorporating industrial emissions into residential and commercial
proportionally.  This was done as the industrial sectors of ACES/Vulcan are
dominated by point sources, many of which don't even rely on natural gas. It
was chosen to split by emissions rather than number of customers as this was
considered more representative (e.g., there may be many residential
customers, yet the commercial ones consume more natural gas overall).

Maintenance and upsets are calculated using national GHGI emission factors
and split into residential and commercial in the same manner as service and
pipeline emissions were.

Post-meter emissions are calculated, strictly for residential, using the
volume of gas delivered to residential customers and the
natural_gas_post_meter_emission_factor, which is based on Fischer et al. by
default.  Fischer et al. measured whole-house emissions from 75 homes in
California using mass balance.

Finally, ACES and/or Vulcan residential and commercial gridded CO2 emission
maps are loaded in.  The emissions for each subsector are then distributed
using the ACES or Vulcan CO2 inventory.  This can be done at the LDC, state,
or domain level.  Emissions at this point are at the state level if not
calculating by LDC and can then be aggregated to the domain level. Otherwise
emissions are at the LDC level and can be aggregated to the state or domain
level.  As such, producing output at the state and LDC level will result in
slightly different output than running only at the state level.

So, to summarize this relatively complex sector, PHMSA data on miles of
pipeline by type and number of services by type of pipeline is combined with
emission factors to calculate emissions.  GHGRP state average numbers of M&R
facilities per mile are also combined with an emission factor, broken down
by pressure.  If calculating byLDC and GHGRP data exists for an LDC, the
reported counts are used instead.  These emissions are then distributed to
the appropriate LDC territories using HIFLD shapefiles, or aggregated at the
state level. To further disaggregate the emissions, they are broken into
residential and commercial portions using the fraction of customers as
reported to the EIA. They are then distributed to the pixel scale using the
residential and commercial sectors of the CO2 inventories Vulcan/ACES.  As
such, GHGRP emissions are not directly used at all, PHMSA is the source of
most activity data with GHGRP providing the numbers of M&R facilities, HIFLD
solely provides shapefiles for each LDC if operating by LDC, and EIA
provides a breakdown of residential vs commercial customers.

GHGRP data is available for this sector starting in 2011 and generally is
about 2 years behind present day, the GHGI is available starting in 1990 and
is updated approximately in sync with the GHGRP.  The HIFLD dataset is
updated infrequently and is available for 2019 and 2017.  The EIA form is
annually reported and is available starting in 1997 and is generally updated
in September to the previous year (i.e., Sept 2024 adds 2023 data).  The
PHMSA data is annual and is available starting in 1970 and is available up
to the most recent year.  The GHGRP includes only facilities that emit at
least 25,000 metric tons of carbon dioxide equivalent while the GHGI is
intended to capture all national emissions. All other datasets are meant to
be inclusive of national facilities. All data is annual.  The GHGI is
national totals while all other datasets are at the facility scale.

GHGRP data, GHGI data, and the HIFLD shapefile will be automatically
downloaded.

The GHGRP is available at \url{https://ghgdata.epa.gov/ghgp/main.do}. For
individual LDC metering and regulating station counts, among other
variables, one must filter to the natural gas local distribution companies
sector, select an individual facility and select "View reported data".  The
GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}.
The necessary GHGI Annex data is available at
\url{https://www.epa.gov/ghgemissions/natural-gas-and-petroleum-systems-ghg-inventory-additional-information-1990-2022-ghg}
for the 2024 GHGI.  In the GHGI Annexes, available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks-1990-2022},
for 2024 there is a link to the file in Section 3.6: "Methodology for
Estimating CH4, CO2, and N2O Emissions from Natural Gas Systems".  The excel
file has multiple sheets, each of which has a separate layout.  The HIFLD
dataset is available at
\url{https://hifld-geoplatform.hub.arcgis.com/datasets/geoplatform::natural-gas-service-territories/about}
and can be donwloaded as both a shapefile or a csv. EIA form 176 is
available at
\url{https://www.eia.gov/naturalgas/ngqs/#?report=RP4&year1=2020&year2=2020&company=Name}
and can be downloaded as an excel file.  The PHMSA Gas Distribution Annual
Data can be download at
\url{https://www.phmsa.dot.gov/data-and-statistics/pipeline/gas-distribution-gas-gathering-gas-transmission-hazardous-liquids}
as a zip file with an excel file for each year. ACES is available at
\url{https://doi.org/10.3334/ORNLDAAC/1943} and Vulcan is available at
\url{https://doi.org/10.3334/ORNLDAAC/1741}.

See references \href{https://doi.org/10.1021/acs.est.0c00437}{Weller et
al.}, \href{https://doi.org/10.1021/acs.est.8b03217}{Fischer et al.},
\href{https://doi.org/10.1029/2020JD032974}{Vulcan} and,
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\references{
\href{https://doi.org/10.1021/acs.est.0c00437}{Weller et al.}

\href{https://doi.org/10.1021/acs.est.8b03217}{Fischer et al.}

\href{https://doi.org/10.1029/2020JD032974}{Vulcan}

\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.

\code{\link[=disaggregation]{disaggregation()}} Disaggregates data to pixels using a sectoral CO2 inventory.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
\keyword{internal}
