% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_individual_sectors.R
\name{prep_plot_data}
\alias{prep_plot_data}
\title{Prepare input data for a logarithmic plot}
\usage{
prep_plot_data(input)
}
\arguments{
\item{input}{SpatRaster.  Intended to be a SpatRaster of gridded methane
emissions.}
}
\value{
The log-scaled input data is returned, after removing infinite
values.
}
\description{
This is a simple helper function for ease of use.  It just takes
the log10 of the data and sets infinite values to NA.
}
\details{
This function is intended to be used when verbose = TRUE for
\code{\link{CH4_inventory_build}}.  It is called by log_plot to help build
plots for each individual sector/subsector.  Some of these, particularly
those for point sources, are more useful on a log scale.
}
\seealso{
\code{\link[=log_plot]{log_plot()}} plots the data in simple log-scale visuals.
}
\keyword{internal}
