% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_GEPA.R
\name{Prepare_GEPA}
\alias{Prepare_GEPA}
\title{Pull gridded methane emissions maps for sectors not built with other
functions from the Maasakkers gridded EPA inventory}
\usage{
Prepare_GEPA(
  inventory_year,
  input_directory,
  output_directory,
  plot_directory,
  County_Tigerlines,
  State_CB,
  domain,
  domain_template,
  verbose
)
}
\arguments{
\item{inventory_year}{Numeric indicating the desired year of data to use.}

\item{input_directory}{Character providing the full filepath to save/load
input data}

\item{output_directory}{Character providing the full filepath to save
processed data}

\item{plot_directory}{Character providing the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{County_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile downloaded in Main.}

\item{State_CB}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{verbose}{Logical indicating whether to save visuals.  It includes 3
plots of the gridded methane emissions, 1 for each netcdf file.}
}
\value{
Nothing is returned from the function, but the main outputs are 3
netcdf files of the methane emissions from the gridded EPA product.  They
are titled "GEPA_thermo.nc" for gridded EPA thermogenic,
"GEPA_non_thermo.nc" for non-thermogenic, "GEPA_ind_landfill.nc" for
industrial landfills.
}
\description{
\code{Prepare_GEPA} writes 3 netcdf files of gridded methane emissions
\itemize{
\item 1 for thermogenic sectors, 1 for industrial landfills, and 1 for
non-thermogenic sources.
}
}
\details{
This function downloads the most appropriate year of gridded methane
data at 0.1 deg by 0.1 deg from Maasakkers et al. and pulls sectors that
have not been calculated in other scripts.  They are projected to the domain
grid and combined into thermogenic and non-thermogenic groups, with the
exception of industrial landfills which are saved separately.  For
thermogenic this includes
\itemize{
\item mobile combustion
\item Coal (abandoned, surface, and underground)
\item Petroleum Systems (exploration, production, refining, and transport)
\item Abandoned Oil and Gas
\item Natural Gas (exploration, processing, and production)
\item Industry (petrochemical, and ferroalloy)
}
And for non-thermogenic this includes
\itemize{
\item Composting
\item Enteric Fermentation
\item Manure Management
\item Rice Cultivation
\item Field Burning
}

The data is available at \url{https://doi.org/10.5281/zenodo.8367082}.  The
closest year will be automatically downloaded.

See reference \href{https://doi.org/10.1021/acs.est.3c05138}{Maasakkers et
al.}
}
\examples{
library(terra)
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)
Prepare_GEPA(inventory_year=2018,
             input_directory="~/../Desktop/in/",
             output_directory="~/../Desktop/out/",
             domain=grid_vect,
             domain_template=grid,
             verbose=TRUE,
             State_CB=vect("~/../Desktop/in/State_CB/tl_2018_us_state.shp"),
             County_Tigerlines=vect("~/../Desktop/in/County_Tigerlines/tl_2018_us_county.shp"),
             plot_directory="~/../Desktop/plots/")
}
\references{
\href{https://doi.org/10.1021/acs.est.3c05138}{Maasakkers et al.}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
