% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetland_emissions_r2.R
\name{SOCCR_Wetlands}
\alias{SOCCR_Wetlands}
\title{Create SOCCR based wetland methane maps}
\usage{
SOCCR_Wetlands(
  input_directory,
  output_directory,
  plot_directory,
  state_name_list,
  domain,
  domain_template,
  Use_SOCCR1,
  Use_SOCCR2,
  Include_freshwater,
  Wetland_EFs,
  verbose,
  County_Tigerlines,
  State_CB,
  Source_Watershed_file,
  Use_NLCD,
  Use_NALCMS,
  Use_Wetcharts,
  Wetcharts_model_subset
)
}
\arguments{
\item{output_directory}{Character providing the full filepath to save
processed data}

\item{plot_directory}{Character providing the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{Use_SOCCR1}{Logical.  Pulled from config file.  Indicating whether or
not to calculate emissions using SOCCR1.}

\item{Use_SOCCR2}{Logical.  Pulled from config file.  Indicating whether or
not to calculate emissions using SOCCR2.}

\item{Include_freshwater}{Logical.  Pulled from config file.  Indicating
whether or not to calculate emissions for freshwater wetlands using
Rosentreter et al.}

\item{Wetland_EFs}{Data frame.  Pulled from config file. Emission factors to
use for all wetlands - including SOCCR1, SOCCR2, and Rosentreter et al.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions on log scales, saved
separately for SOCCR1, SOCCR2, and freshwater.}

\item{County_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{State_CB}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{Watershed_file}{Character.  Commission for Environmental
Cooperation watershed shapefile.  Available at
\url{http://www.cec.org/north-american-environmental-atlas/watersheds/}.
Only relevant if USE_SOCCR2 = TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
3 netcdcf files of the methane emissions from wetlands with 1 file for
SOCCR1 and 1 file for SOCCR2 based emissions.  Lakes and rivers from
Rosentreter et al. emissions can also be included.  They are titled
"SOCCR1.nc", "SOCCR2.nc", and "Freshwater.nc".

If verbose is set to TRUE, then multiple figures are also saved.  Log scale
plots with consistent axes are saved for the 2 SOCCR emissions and
freshwater emissions.  They are saved as "SOCCR1.png", "SOCCR2.png", and
"Freshwater.png".
}
\description{
\code{SOCCR_Wetlands} writes up to 3 netcdf files of gridded wetland
methane emissions, 1 for SOCCR1, 1 for SOCCR2, and 1 for freshwater.
Includes optional visuals as well.
}
\details{
This function takes the output of \code{NWI_Wetland_fraction}, which is
per pixel fractional coverage of wetlands separated by wetland type, and
applies emission factors to convert coverage to methane emissions.  It is
simply applying SOCCR1 or SOCCR2 average emissions from wetlands to the NWI
activity data.  Additionally, freshwater emissions can be calculated using
emissions from Rosentreter et al.

SOCCR1 values are based on the arithmetic averages of Table F5, SOCC2 values
are based on the arithmetic averages of Tables 13B.8 to 13B.11 for PFO and
PNF and Table 15A.2 for M2 and E2, and Lakes and Rivers (L1, L2, and R1 -
R4) are from Rosentreter et al. using the median flux from rivers and the
largest lake class (>1 km).  SOCCR2 is calculated by watershed as there was
regionally separated data.  The lake flux was chosen as McDonald et al. show
that large lakes (>1 km2) constitute 71\\% of the total lake area in CONUS
(rising to 90\\% if including the Great Lakes).

SOCCR1 is available at
\url{https://www.carboncyclescience.us/state-carbon-cycle-report-soccr} and
SOCCR2 is available at \url{https://carbon2018.globalchange.gov/}.

See references \href{https://doi.org/10.4319/lo.2012.57.2.0597}{McDonald et
al.} and \href{https://doi.org/10.1038/s41561-021-00715-2}{Rosentreter et al.}
}
\examples{
library(terra)
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)
EFs <- data.frame("E2"=c(10.3,15.29*16.043/12.011),
                          "M2"=c(10.3,15.29*16.043/12.011),
                          "PFO"=c(36,18.52*16.043/12.011),
                          "PNF"=c(36,24.92*16.043/12.011),
                          "L1"=5,
                          "L2"=5,
                          "R1"=7.88,
                          "R2"=7.88,
                          "R3"=7.88,
                          "R4"=7.88)
rownames(EFs) <- c("SOCCR1","SOCCR2")

# convert from g CH4 per m2 per yr to nmol/m2/s
EFs=EFs*1E9/(16.043*365.25*24*60*60)

SOCCR_Wetlands(output_directory="~/../Desktop/out/",
               plot_directory="~/../Desktop/plots/",
               domain=grid,
               Use_SOCCR1=TRUE,
               Use_SOCCR2=TRUE,
               Include_freshwater=TRUE,
               Wetland_EFs=EFs,
               verbose=TRUE,
               County_Tigerlines=vect("~/../Desktop/in/County_Tigerlines/tl_2018_us_county.shp"),
               State_CB=vect("~/../Desktop/in/State_CB/tl_2018_us_state.shp"),
               Watershed_file="~/../Desktop/in/watersheds_shapefile/watershed_p_v2.shp")

}
\references{
\href{https://doi.org/10.4319/lo.2012.57.2.0597}{McDonald et al.}

\href{https://doi.org/10.1038/s41561-021-00715-2}{Rosentreter et
al.}
}
\seealso{
\itemize{
\item \code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
\item \code{\link[=Disaggregate_Wetcharts]{Disaggregate_Wetcharts()}} Calculates methane emissions for the wetland sector using wetcharts instead.
\item \code{\link[=NWI_Wetland_fraction]{NWI_Wetland_fraction()}} Calculates the fraction of wetland land cover by wetland type in each pixel.
}
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
