% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetland_emissions_r2.R
\name{SOCCR_Wetlands}
\alias{SOCCR_Wetlands}
\title{Create SOCCR based wetland methane maps}
\usage{
SOCCR_Wetlands(
  input_directory,
  output_directory,
  plot_directory,
  state_name_list,
  domain,
  domain_template,
  Use_SOCCR1,
  Use_SOCCR2,
  Wetland_EFs,
  verbose,
  County_Tigerlines,
  State_CB,
  Source_Watershed_file,
  Use_Wetcharts,
  Wetcharts_model_subset
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{Use_SOCCR1}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Use_SOCCR2}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Wetland_EFs}{Data.frame.  Pulled from \code{\link{M3T_config}}.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions on log scales, saved
separately for SOCCR1, SOCCR2, and freshwater.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}

\item{Source_Watershed_file}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Use_Wetcharts}{Logical.  Pulled from \code{\link{M3T_config}}.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
3 netcdcf files of the methane emissions from wetlands with 1 file for
SOCCR1 based emissions, 1 file for SOCCR2 based emissions, and a file for
lakes and rivers.  They are titled "SOCCR1.nc", "SOCCR2.nc", and
"Freshwater.nc".

If verbose is set to TRUE, then multiple figures are also saved.  Log scale
plots with consistent axes are saved for the 2 SOCCR emissions and
freshwater emissions.  They are saved as "SOCCR1.png", "SOCCR2.png", and
"Freshwater.png".
}
\description{
\code{SOCCR_Wetlands} writes up to 3 netcdf files of gridded wetland
methane emissions, 1 for SOCCR1, 1 for SOCCR2, and 1 for freshwater.
Includes optional visuals as well.
}
\details{
This function takes the output of \code{NWI_Wetland_fraction}, which is
per pixel fractional coverage of wetlands separated by wetland type, and
applies emission factors to convert coverage to methane emissions.  It is
simply applying SOCCR1 or SOCCR2 average emissions from wetlands to the NWI
activity data.  Additionally, freshwater emissions can be calculated using
emissions from Rosentreter et al.

SOCCR1 values are based on the arithmetic averages of Table F5, SOCC2 values
are based on the arithmetic averages of Tables 13B.8 to 13B.11 for PFO and
PNF and Table 15A.2 for M2 and E2, and Lakes and Rivers (L1, L2, and R1 -
R4) are from Rosentreter et al. using the median flux from rivers and the
largest lake class (>1 km).  SOCCR2 is calculated by watershed as there was
regionally separated data.  The lake flux was chosen as McDonald et al. show
that large lakes (>1 km2) constitute 71\\% of the total lake area in CONUS
(rising to 90\\% if including the Great Lakes).

SOCCR1 is available at
\url{https://www.carboncyclescience.us/state-carbon-cycle-report-soccr} and
SOCCR2 is available at \url{https://carbon2018.globalchange.gov/}.

See references \href{https://doi.org/10.4319/lo.2012.57.2.0597}{McDonald et
al.} and \href{https://doi.org/10.1038/s41561-021-00715-2}{Rosentreter et al.}
}
\references{
\href{https://doi.org/10.4319/lo.2012.57.2.0597}{McDonald et al.}

\href{https://doi.org/10.1038/s41561-021-00715-2}{Rosentreter et
al.}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.

\code{\link[=Disaggregate_Wetcharts]{Disaggregate_Wetcharts()}} Calculates methane emissions for the wetland
sector using wetcharts instead.

\code{\link[=NWI_Wetland_fraction]{NWI_Wetland_fraction()}} Calculates the fraction of wetland land cover by
wetland type in each pixel.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
\keyword{internal}
