% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_combustion_r4.R
\name{Stationary_combustion}
\alias{Stationary_combustion}
\title{Create gridded stationary combustion methane emissions maps}
\usage{
Stationary_combustion(
  input_directory,
  domain,
  domain_template,
  state_name_list,
  output_directory,
  inventory_year,
  verbose,
  County_Tigerlines,
  Use_ACES,
  Use_Vulcan,
  aces_res,
  aces_com,
  aces_ind,
  aces_elec,
  vu_res,
  vu_com,
  vu_ind,
  vu_elec,
  stationary_combustion_by_state,
  stationary_combustion_by_domain,
  stationary_combustion_GHGI_data,
  stationary_combustion_emission_factors,
  Source_EIA_SEDS_data,
  Source_NEI_data,
  EIA_API_key,
  plot_directory,
  State_CB
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
raw input data}

\item{domain}{SpatVector polygon outlining the desired output area}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system}

\item{state_name_list}{Character vector listing all states within the desired
domain}

\item{output_directory}{Character providing the full filepath to save
processed data}

\item{inventory_year}{Character indicating the desired year of data to use.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions for each
fuel-sector-inventory-variation combination as well as 2 summed plots for
each inventory-variation combination - one for wood and one for all other
sectors.}

\item{County_Tigerlines}{SpatVector.  United States Census Bureau county
shapefile downloaded in Main.}

\item{Use_ACES}{Logical. Pulled from config file. indicating whether or not to
use ACES to disaggregate from county-level to pixel level emissions.  Either
ACES or Vulcan must be used, though both can be.}

\item{Use_Vulcan}{Logical. Pulled from config file. indicating whether or not
to use Vulcan to disaggregate from county-level to pixel level emissions.
Either ACES or Vulcan must be used, though both can be.}

\item{stationary_combustion_by_state}{Logical. Pulled from config file.
indicating whether state-toal emissions should be distributed to the county
scale as is.  Either bystate or bydomain must be used, though both can be.}

\item{stationary_combustion_by_domain}{Logical. Pulled from config file.
indicating whether state-toal emissions should be aggregated to the domain
and then distributed to the county scale.  Either bystate or bydomain must
be used, though both can be.}

\item{stationary_combustion_GHGI_data}{Data frame.  Pulled from config file. 1
by 15 data frame with consumption for each sector-fuel combination from the
GHGI. Although the data is national, there is a state entry set to US_EPA.
Consumption is in thousands of British Thermal Units (BTU).  The GHGI is
available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}
and the values can be found in the Annexes in a table titled "Fuel
Consumption by Stationary Combustion for Calculating CH4 and N2O Emissions
(TBtu)".  Includes every combination of residential (res), commercial (com),
electric (elec), and residential (res) sectors with coal, petroleum (petr),
natural gas (gas), and wood fuels with names using the abbreviations in
parentheses (e.g., com_coal, elec_petr).  Excludes res_coal and res_gas as
residential coal does not exist in the U.S. anymore and residential gas is
accounted for elsewhere.}

\item{stationary_combustion_emission_factors}{Data frame.  Pulled from config
file.  1 by 14 data frame with emission factors for each sector-fuel
combination from the IPCC.  Built equivalently to the GHGI_data, but without
an entry for the state.  Emission factors are in g/GJ, equivalent to kg/TJ.
Default emission factors are available in IPCC 2006 volume 2: Energy, tables
2.2 through 2.5
\url{https://www.ipcc-nggip.iges.or.jp/public/2006gl/vol2.html}. The natural
gas electric sector emission factor is instead pulled from Hajny et al.,
2019 \url{https://doi.org/10.1021/acs.est.9b01875}.  This is 5.7 kg/TJ,
within uncertainties of the GHGI value of 4.1 kg/TJ, both of which are
larger than the IPCC default of 1 kg/TJ.}

\item{EIA_API_key}{Character.  Pulled from config file.  API key to access
SEDS data API.  The API is described at \url{https://www.eia.gov/opendata/}
and one can register for a key with a link on the right hand side of this
page.}

\item{plot_directory}{Character providing the full filepath to save figures.
Only relevant if verbose = TRUE.}

\item{State_CB}{SpatVector.  United States Census Bureau county
shapefile.  Available at
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}.
Only relevant if verbose=TRUE.}

\item{ACES_directory}{Character providing the full path to a folder containing
the ACES sectoral CO2 inventories.  Must include the residential,
commercial, electric (elec), and industrial sectors.  ACES v2.0 is available
at \url{https://doi.org/10.3334/ORNLDAAC/1943}, though the hourly file
should be averaged across hours to create an annually averaged inventory.
There is a function to do this as well, though it is quite time consuming
given the amount of data that needs to be downloaded.  The year closest to
"inventory_year" is used, but those further from that year are considered if
the closest is unavailable.}

\item{vulcan_directory}{Character providing the full path to a folder
containing the Vulcan sectoral CO2 inventories.  Must include the
residential, commercial, electric (elec_prod), and industrial sectors.
Vulcan v3.0 is available at \url{https://doi.org/10.3334/ORNLDAAC/1741}, and
the annual mean files should be used.  The year closest to "inventory_year"
is used.  As all years are contained in the same file, it does not search
for other years.}

\item{ACES_year}{Numeric providing the year of ACES data to use, calculated in
Main.}

\item{vulcan_band}{Numeric providing the band of Vulcan data to use (1-6 =
2010 - 2015), calculated in Main.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
56 netcdf files of the methane emissions from stationary combustion.  They
are titled as "stat_comb_sector_fuel_variation_inventory.nc" where sector is
abbreviated as com (commercial), res (residential), elec (electric), and ind
(industrial); fuel is abbreviated as wood, petr (petroleum), gas (natural
gas), and coal; variation is bystate or bydomain; and inventory is ACES
or Vulcan.
}
\description{
\code{Stationary_combustion} writes up to 56 netcdf files of gridded
methane emissions from stationary combustion sources, as well as optional
visuals
}
\details{
This function calculates and grids methane emissions from stationary
combustion. It uses the Energy Information Administration's (EIA) State
Energy Data System (SEDS), Environmental Protection Agency's (EPA)
Greenhouse Gas Inventory (GHGI), EPA National Emissions Inventory (NEI) and
the Vulcan and/or Anthropogenic Carbon Emission System (ACES) CO2
inventories.  First, the ratio of SEDS national consumption by fuel-sector
combination to the equivalent GHGI data is applied to SEDS state data so
that they are approximately consistent with the GHGI.  The state total
consumption is then converted to emissions using emission factors.  There
are two variations at this step.

bystate: The state level emissions are then distributed to the county level
using NEI CO concentrations to calculate weighting values for each county
(i.e., county CO / state total CO).

bydomain: The state level emissions are summed to get a total for all states
in the domain.  This is distributed to the county level using NEI CO
concentrations to calculate weighting values for each county (i.e., county
CO / domain total CO).

Lastly, county total emissions are distributed using the ACES and/or Vulcan
CO2 inventories.

This entire process is done separately for each fuel and sector using the
GHGI, SEDS, and NEI data that are broken down by sector and fuel and CO2
inventories that are broken down by sector.  Sectors considered are
residential, commercial, industrial, and electric and the fuels considered
are coal, natural gas, petroleum, and wood.  Residential coal is ignored as
it does not exist in the U.S. anymore and residential gas is ignored as it
is accounted for elsewhere.

The necessary SEDS and NEI data will be automatically downloaded.

The GHGI is intended to capture all national emissions.  The SEDS
consumption data provides state-level consumption broken down by fuel (coal,
natural gas, petroleum, wood, geothermal, solar, and electricity) and sector
(residential, commercial, industrial, electric, transportation).  The NEI
data being used provides CO emissions at the county level.  The GHGI is
available starting in 1990 and is generally about 2 years behind present
day.  SEDS data is available starting in 1960 and generally is about 2 years
behind present day.  NEI data is available beginning in at least 1990, is
released every three years, and generally takes three years to complete
(i.e., 2023 NEI is released in 2026).  All data is annual.  The SEDS data is
at the state scale, NEI data is at the county scale, and GHGI data is at the
national scale.

The GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}
The SEDS is available at
\url{https://www.eia.gov/state/seds/seds-data-complete.php?sid=US} The NEI
is available at
\url{https://www.epa.gov/air-emissions-inventories/national-emissions-inventory-nei}
ACES is available at \url{https://doi.org/10.3334/ORNLDAAC/1943} and Vulcan
is available at \url{https://doi.org/10.3334/ORNLDAAC/1741}.

Each fuel-sector-inventory-variation combination is saved separately.

See references
\href{https://doi.org/10.1029/2020JD032974}{Vulcan} and
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\examples{
library(terra)
user_key = "__user_EIA_API_key__"
grid_bbox=cbind(c(-76.65,-73.65),c(38.97,40.97))
grid_res=0.01
grid_crs="epsg:4326"
grid <- rast(nrows=diff(range(grid_bbox[,2]))/grid_res,
             ncols=diff(range(grid_bbox[,1]))/grid_res, xmin=min(grid_bbox[,1]),
             xmax=max(grid_bbox[,1]), ymin=min(grid_bbox[,2]), ymax=max(grid_bbox[,2]),
             crs=grid_crs)
grid_vect <- as.polygons(ext(grid),crs=grid_crs)
Stationary_combustion(domain=grid_vect,
                      domain_template=grid,
                      state_name_list=c("DE","MD","NJ","NY","PA"),
                      input_directory="~/../Desktop/in/",
                      output_directory="~/../Desktop/out/",
                      inventory_year=2018,
                      verbose=TRUE,
                      Use_ACES=TRUE,
                      Use_Vulcan=TRUE,
                      ACES_directory="~/../Desktop/in/Inventories/ACES_v2.0",
                      vulcan_directory="~/../Desktop/in/Inventories/Vulcan_v3.0",
                      ACES_year=2017,
                      vulcan_band=6,
                      stationary_combustion_by_state=TRUE,
                      stationary_combustion_by_domain=TRUE,
                      stationary_combustion_GHGI_data=data.frame(
                        "State"="US_EPA",
                        "com_coal"=17,
                        "ind_coal"=517,
                        "elec_coal"=10554,
                        "res_petr"=975,
                        "com_petr"=801,
                        "ind_petr"=2062,
                        "elec_petr"=42,
                        "com_gas"=3647,
                        "ind_gas"=9484,
                        "elec_gas"=11553,
                        "res_wood"=544,
                        "com_wood"=84,
                        "ind_wood"=1407,
                        "elec_wood"=68),
                      stationary_combustion_emission_factors=data.frame(
                        "com_coal"=10,
                        "ind_coal"=10,
                        "elec_coal"=1,
                        "res_petr"=10,
                        "com_petr"=10,
                        "ind_petr"=3,
                        "elec_petr"=3,
                        "com_gas"=5,
                        "ind_gas"=1,
                        "elec_gas"=5.4/(1.0550559*0.9), #g/mmbtu to g/GJ and low to high heating value (0.9)
                        "res_wood"=300,
                        "com_wood"=300,
                        "ind_wood"=30,
                        "elec_wood"=30),
                      EIA_API_key=user_key,
                      State_CB=vect("~/../Desktop/in/State_CB/tl_2018_us_state.shp"),
                      County_Tigerlines=vect("~/../Desktop/in/County_Tigerlines/tl_2018_us_county.shp"),
                      plot_directory="~/../Desktop/plots/")
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings provided in config.
}
\author{
Joe Pitt, \email{madeup@wisc.edu}

Kris Hajny, \email{blank@fake.edu}

Israel Lopez-Coto, \email{test@test.edu}
}
