% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stationary_combustion.R
\name{Stationary_combustion}
\alias{Stationary_combustion}
\title{Create gridded stationary combustion methane emissions maps}
\usage{
Stationary_combustion(
  input_directory,
  domain,
  domain_template,
  state_name_list,
  output_directory,
  inventory_year,
  GHGI_data_yr,
  State_Tigerlines,
  verbose,
  County_Tigerlines,
  Use_ACES,
  Use_Vulcan,
  aces_res,
  aces_com,
  aces_ind,
  aces_elec,
  vu_res,
  vu_com,
  vu_ind,
  vu_elec,
  stationary_combustion_by_state,
  stationary_combustion_by_domain,
  stationary_combustion_GHGI_data,
  stationary_combustion_emission_factors,
  Source_EIA_SEDS_data,
  Source_NEI_data,
  EIA_API_key,
  plot_directory,
  State_CB
)
}
\arguments{
\item{input_directory}{Character providing the full filepath to save/load
input data.  Automatically defined in \code{\link{CH4_inventory_build}} from
the input \code{run_directory}.}

\item{domain}{SpatVector polygon outlining the desired output area as created
in \code{\link{CH4_inventory_build}}.}

\item{domain_template}{SpatRaster providing the desired output grid, including
the desired resolution and coordinate reference system as created in
\code{\link{CH4_inventory_build}}.}

\item{state_name_list}{Character vector listing all states within the desired
domain as created in \code{\link{CH4_inventory_build}}.}

\item{output_directory}{Character providing the full filepath to save
processed data as created in \code{\link{CH4_inventory_build}}}

\item{inventory_year}{Numeric indicating the desired year of data to use.  The
closest available will be used if unavailable with a user update.}

\item{GHGI_data_yr}{Integer providing the year of data to use for the
\href{https://www.epa.gov/ghgreporting}{Environmental Protection Agency
(EPA) Greenhouse Gas Reporting Program (GHGRP) data}.}

\item{verbose}{Logical indicating whether to save additional output.  This
includes plots of the gridded methane emissions for each
fuel-sector-inventory-variation combination as well as 2 summed plots for
each inventory-variation combination - one for wood and one for all other
sectors.}

\item{County_Tigerlines}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{US
Census Tigerlines} files for visualization. Only relevant if \code{verbose} = TRUE.}

\item{Use_ACES}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{Use_Vulcan}{Logical.  Pulled from \code{\link{M3T_config}}.}

\item{aces_res}{SpatRaster of residential CO2 emissions from the ACES
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_ACES} and \code{Source_ACES}.}

\item{aces_com}{SpatRaster of commercial CO2 emissions from the ACES
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_ACES} and \code{Source_ACES}.}

\item{aces_ind}{SpatRaster of industrial CO2 emissions from the ACES inventory
as loaded in \code{\link{CH4_inventory_build}} based on \code{Use_ACES} and
\code{Source_ACES}.}

\item{aces_elec}{SpatRaster of electric power production CO2 emissions from
the ACES inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_ACES} and \code{Source_ACES}.}

\item{vu_res}{SpatRaster of residential CO2 emissions from the Vulcan v4.0
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{vu_com}{SpatRaster of commercial CO2 emissions from the Vulcan v4.0
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{vu_ind}{SpatRaster of industrial CO2 emissions from the Vulcan v4.0
inventory as loaded in \code{\link{CH4_inventory_build}} based on
\code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{vu_elec}{SpatRaster of electric power production CO2 emissions from the
Vulcan v4.0 inventory as loaded in \code{\link{CH4_inventory_build}} based
on \code{Use_Vulcan} and \code{Source_Vulcan}.}

\item{stationary_combustion_by_state}{Logical.  Pulled from
\code{\link{M3T_config}}.}

\item{stationary_combustion_by_domain}{Logical.  Pulled from
\code{\link{M3T_config}}.}

\item{stationary_combustion_GHGI_data}{Data frame.  Pulled from
\code{\link{M3T_config}}.}

\item{stationary_combustion_emission_factors}{Data frame.  Pulled from
\code{\link{M3T_config}}.}

\item{Source_EIA_SEDS_data}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{Source_NEI_data}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{EIA_API_key}{Character.  Pulled from \code{\link{M3T_config}}.}

\item{plot_directory}{Character. \strong{Optional}. Provides the full filepath
to save figures. Only relevant if \code{verbose} = TRUE.}

\item{State_CB}{SpatVector. \strong{Optional}.
\href{https://www.census.gov/geographies/mapping-files/tme-series/geo/cartographic-boundary.html}{US
Census Cartographic Boundary} files for visualization. Only relevant if
\code{verbose} = TRUE.}
}
\value{
Nothing is returned from the function, but the main outputs are up to
56 netcdf files of the methane emissions from stationary combustion.  They
are titled as "stat_comb_sector_fuel_variation_inventory.nc" where sector is
abbreviated as com (commercial), res (residential), elec (electric), and ind
(industrial); fuel is abbreviated as wood, petr (petroleum), gas (natural
gas), and coal; variation is bystate or bydomain; and inventory is ACES or
Vulcan.
}
\description{
\code{Stationary_combustion} writes up to 56 netcdf files of gridded
methane emissions from stationary combustion sources, as well as optional
visuals
}
\details{
This function calculates and grids methane emissions from stationary
combustion. It uses the Energy Information Administration's (EIA) State
Energy Data System (SEDS), Environmental Protection Agency's (EPA)
Greenhouse Gas Inventory (GHGI), EPA National Emissions Inventory (NEI) and
the Vulcan and/or Anthropogenic Carbon Emission System (ACES) CO2
inventories.  First, the ratio of SEDS national consumption by fuel-sector
combination to the equivalent GHGI data is applied to SEDS state data so
that they are approximately consistent with the GHGI.  The state total
consumption is then converted to emissions using emission factors.  There
are two variations at this step.

bystate: The state level emissions are then distributed to the county level
using NEI CO concentrations to calculate weighting values for each county
(i.e., county CO / state total CO).

bydomain: The state level emissions are summed to get a total for all states
in the domain.  This is distributed to the county level using NEI CO
concentrations to calculate weighting values for each county (i.e., county
CO / domain total CO).

Lastly, county total emissions are distributed using the ACES and/or Vulcan
CO2 inventories.

This entire process is done separately for each fuel and sector using the
GHGI, SEDS, and NEI data that are broken down by sector and fuel and CO2
inventories that are broken down by sector.  Sectors considered are
residential, commercial, industrial, and electric and the fuels considered
are coal, natural gas, petroleum, and wood.  Residential coal is ignored as
it does not exist in the U.S. anymore and residential gas is ignored as it
is accounted for elsewhere.

The necessary SEDS and NEI data will be automatically downloaded.

The GHGI is intended to capture all national emissions.  The SEDS
consumption data provides state-level consumption broken down by fuel (coal,
natural gas, petroleum, wood, geothermal, solar, and electricity) and sector
(residential, commercial, industrial, electric, transportation).  The NEI
data being used provides CO emissions at the county level.  The GHGI is
available starting in 1990 and is generally about 2 years behind present
day.  SEDS data is available starting in 1960 and generally is about 2 years
behind present day.  NEI data is available beginning in at least 1990, is
released every three years, and generally takes three years to complete
(i.e., 2023 NEI is released in 2026).  All data is annual.  The SEDS data is
at the state scale, NEI data is at the county scale, and GHGI data is at the
national scale.

The GHGI is available at
\url{https://www.epa.gov/ghgemissions/inventory-us-greenhouse-gas-emissions-and-sinks}
The SEDS is available at
\url{https://www.eia.gov/state/seds/seds-data-complete.php?sid=US} The NEI
is available at
\url{https://www.epa.gov/air-emissions-inventories/national-emissions-inventory-nei}
ACES is available at \url{https://doi.org/10.3334/ORNLDAAC/1943} and Vulcan
is available at \url{https://doi.org/10.3334/ORNLDAAC/1741}.

Each fuel-sector-inventory-variation combination is saved separately.

See references \href{https://doi.org/10.1029/2020JD032974}{Vulcan} and
\href{https://doi.org/10.1002/2017JD027359}{ACES}
}
\seealso{
\code{\link[=CH4_inventory_build]{CH4_inventory_build()}} Calculates methane inventory using settings
provided in config.

\link{M3T_config} Generates the config function with user-editable settings used
throughout processing.

\code{\link[=Inventory_based_disaggregation]{Inventory_based_disaggregation()}} Disaggregates data from the county level to pixels using a
sectoral CO2 inventory.
}
\keyword{internal}
