% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Trycatch_downloader}
\alias{Trycatch_downloader}
\title{Helper function for downloading data}
\usage{
Trycatch_downloader(URL, output_location = NULL, method, error_message = "")
}
\arguments{
\item{URL}{Character. The URL for the data}

\item{output_location}{Character.  The path and filename for the data to be
saved to, if it is being saved.}

\item{method}{Character.  One of save, JSON, or vect.  Save will save it as a
file using output_location, JSON will direct load it into R using the
jsonlite package, and vect will directly load it into R using terra's vect
function (points or polygons, e.g., .shp or .gpkg)}

\item{error_message}{Character.  Message to output should download fail.}
}
\description{
Helper function for downloading data
}
\details{
This wrapper will use the proper function to download the data and
uses a repeat and trycatch loop in case there is poor internet
connectivity.  It will attempt the download up to 5 times before failing.
Includes user updates throughout as well.
}
\author{
Kris Hajny, \email{kris.hajny@gmail.com}

Joe Pitt, \email{joseph.pitt@bristol.ac.uk}
}
\keyword{internal}
