% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeCDF_no_newline}
\alias{writeCDF_no_newline}
\title{Wrapper around terra's writeCDF to prevent newline.}
\usage{
writeCDF_no_newline(input_raster, ...)
}
\arguments{
\item{input_raster}{SpatRaster.}
}
\description{
\code{writeCDF_no_newline} runs terra's writeCDF in a way that avoids
outputting a new line after running.  Improves flow of periodic user
updates within the package.
}
\keyword{internal}
